### 2025-06-26 6f05b32
Gmsh: Sanity check if the mesh contains volume elements after running FLTK GUI
