### 2025-06-26 78ebc23
Re-set the order for newly created elements
