+++ pyhope/mesh/mesh_builtin.py
@@ -334,6 +334,10 @@ def MeshCartesian() -> meshio.Mesh:
     if debugvisu and IsDisplay():
         gmsh.fltk.run()
 
+    # Re-set the order for newly created elements
+    gmsh.model.mesh.setOrder(mesh_vars.nGeo)
+    gmsh.model.geo.synchronize()
+
     # Sanity check if the mesh contains volume elements
     # > User might have modified the mesh inside the FLTK GUI
     gmsh_elems = np.asarray((gmsh.option.getNumber('Mesh.NbTetrahedra'),
