### 2025-06-27 b871098
CI/CD: Pre-install Gmsh instead of removing and re-installing
Co-authored-by: Daniel Appel <daniel.appel@iag.uni-stuttgart.de>
