### 2025-06-27 3c6d284
Docker: Reduce image size
