### 2025-06-27 e9bd365
Docker: Reduce image size, remove FLEXI source
