+++ .docker/python_3.10.16-bookworm/dockerfile
@@ -4,20 +4,33 @@ FROM python:3.10.16-bookworm
 # Numerics research group
 LABEL maintainer="numerics@iag.uni-stuttgart.de"
 
-# Setup required packages
-RUN apt-get update && \
-    apt-get install -y libglu1-mesa-dev freeglut3-dev mesa-common-dev libxcursor1 libxinerama1 && \
-    apt-get clean
-
-# Setup uv package manager
-RUN mkdir -p /.local/bin && \
-    curl -L -o uv.tar.gz https://github.com/astral-sh/uv/releases/download/0.5.13/uv-x86_64-unknown-linux-gnu.tar.gz && \
-    tar -xzf uv.tar.gz -C /.local/bin --strip-components=1 && \
-    rm uv.tar.gz
-
 # Setup uv environment variables
-ENV PATH="/.local/bin:$PATH"
-ENV UV_LINK_MODE=copy
+# > Avoid interactive prompts and reduce extra files
+ENV DEBIAN_FRONTEND=noninteractive \
+    PATH="/.local/bin:$PATH"       \
+    UV_LINK_MODE=copy
 
 # Set architecture argument
 ARG ARCH
+
+# Setup required packages
+RUN apt-get update              && \
+    apt-get install -y --no-install-recommends \
+      libglu1-mesa                 \
+      libglut3.12                  \
+      libxcursor1                  \
+      libxinerama1              && \
+    rm -rf /var/lib/apt/lists/* && \
+    apt-get clean               && \
+    rm -rf                         \
+      /usr/share/doc/*             \
+      /usr/share/man/*             \
+      /var/log/*                   \
+      /tmp/*
+
+# Setup uv package manager
+RUN mkdir -p /.local/bin        && \
+    curl -fsSL -o /tmp/uv.tar.gz   \
+      https://github.com/astral-sh/uv/releases/download/0.7.15/uv-x86_64-unknown-linux-gnu.tar.gz && \
+    tar -xzf /tmp/uv.tar.gz -C /.local/bin --strip-components=1 && \
+    rm /tmp/uv.tar.gz
