### 2025-07-08 c8592e0
PyPI: Update repository/documentation URL
