+++ pyproject.toml
@@ -36,7 +36,8 @@ classifiers = [
 
 [project.urls]
 homepage = "https://numericsresearchgroup.org"
-repository = "https://gitlab.iag.uni-stuttgart.de/flexi/codes/pyhope"
+repository = "https://github.com/hopr-framework/pyhope"
+documentation = "https://hopr.readthedocs.io"
 
 [project.scripts]
 pyhope = "pyhope.script.pyhope_cli:main"
