+++ .gmsh/build_manylinux.py
@@ -28,7 +28,7 @@
 import os
 import sys
 import build
-import errno
+# import errno
 import multiprocessing
 import platform
 import shutil
@@ -100,6 +100,10 @@ def configure(configure_cmd: list, cwd: Optional[str] = None, env: Optional[dict
         print_step('Configuring with: {}'.format(configure_cmd))
         subprocess.run(configure_cmd, check=True, cwd=cwd, env=env, stderr=sys.stderr, stdout=sys.stdout)
 
+    else:
+        print('Unknown configure command')
+        sys.exit(1)
+
 
 def compile(install_cmd: Optional[list] = None, ncores: int = 1, cwd: Optional[str] = None, env: Optional[dict] = None) -> None:
     # Build the software
@@ -110,6 +114,7 @@ def compile(install_cmd: Optional[list] = None, ncores: int = 1, cwd: Optional[s
     if install_cmd:
         print_step('Installing...')
         subprocess.run(install_cmd, check=True, cwd=cwd, stderr=sys.stderr, stdout=sys.stdout)
+
 
 # Get systemc architecture
 arch = platform.machine()
@@ -122,23 +127,23 @@ Doxygen_DIR  = os.path.join(WORK_DIR, 'doxygen')
 
 # HDF5
 HDF5_DIR     = os.path.join(WORK_DIR, 'hdf5')
-HDF5_VERSION = '1.14.5'
+HDF5_VERSION = '1.14.6'
 
 # TK
 TK_DIR       = os.path.join(WORK_DIR, 'tk')
-TK_VERSION = '8.6.15'
+TK_VERSION = '8.6.16'
 
 # TCL
 TCL_DIR      = os.path.join(WORK_DIR, 'tcl')
-TCL_VERSION = '8.6.14'
+TCL_VERSION = '8.6.16'
 
 # CGNS
 CGNS_DIR     = os.path.join(WORK_DIR, 'CGNS')
-CGNS_VERSION = 'v4.4.0'
+CGNS_VERSION = 'v4.5.0'
 
 # OpenCASCADE
 OCC_DIR      = os.path.join(WORK_DIR, 'occt')
-OCC_VERSION  = 'V7_8_1'
+OCC_VERSION  = 'V7_9_1'
 
 # FreeType
 FREETYPE_VERSION = '2.13.3'
@@ -146,37 +151,34 @@ FREETYPE_DIR = os.path.join(WORK_DIR, 'freetype')
 
 # libXFT
 LIBXFT_DIR     = os.path.join(WORK_DIR, 'libxft')
-LIBXFT_VERSION = '2.3.8'
+LIBXFT_VERSION = '2.3.9'
 
 # libPNG
 LIBPNG_DIR     = os.path.join(WORK_DIR, 'libpng')
-LIBPNG_VERSION = '1.6.44'
+LIBPNG_VERSION = '1.6.50'
 
 # libJPEG
 LIBJPEG_TURBO_DIR     = os.path.join(WORK_DIR, 'libjpeg-turbo')
-LIBJPEG_TURBO_VERSION = '3.0.4'
+LIBJPEG_TURBO_VERSION = '3.1.1'
 
 # FLTK
 FLTK_DIR     = os.path.join(WORK_DIR, 'fltk')
-FLTK_VERSION = '1.3.9'
+FLTK_VERSION = '1.3.11'
 # FLTK_VERSION = '1.4.0'  # Currently broken, produces segfault
 
 # gperf
 GPERF_VERSION = "3.1"
 
 # Fontconfig
-FONTCONFIG_VERSION = '2.14.2'
[…]
