+++ pyhope/gmsh/gmsh_install.py
@@ -159,7 +159,7 @@ def PkgsInstallGmsh(system: str, arch: str, version: str) -> None:
     if version == 'nrg':
         # Gitlab "python-gmsh" access
         lfs = 'yes'
-        lib = 'gmsh-{}-py3-none-{}_{}.whl'.format(Gitlab.LIB_VERSION, system, arch)
+        lib = 'gmsh-{}-py3-none-{}_{}.whl'.format(Gitlab.LIB_VERSION[system][arch], system, arch)
 
         # Create a temporary directory
         with tempfile.TemporaryDirectory() as path:
