### 2025-07-24 aaa16b5
Zones: Implement mesh zones for built-in Gmsh generator
Map volume cell sets to elements and pass the info into the HDF5 output
