### 2025-07-24 5a9bb24
Zones: Implement mesh zones for HOPR/Gambit reader
