### 2025-08-07 #8484 @jwallwork23
> Hi @jwallwork23 do you have time to review this paper?

Sure, I can review this. I'm on vacation at the minute but should have time to do it in a couple of weeks.
