### 2025-08-08 #8769 @editorialbot
[REVIEW]: PyHOPE: A Python Toolkit for Three-Dimensional Unstructured High-Order Meshes
**Submitting author:** @kopperp (<a href="http://orcid.org/0000-0002-7613-0739">Patrick Kopper</a>)
**Repository:** https://github.com/hopr-framework/PyHOPE
**Branch with paper.md** (empty if default branch): paper.joss.2025
**Version:** v0.8.0
