### 2025-08-11 1e25d39
Mortars: Add sanity check for edges calculation
