### 2025-08-11 c71814e
XDMF: Write initial XDMF output
Currently, only NGeo=1. Mixed grids are untested
