+++ pyhope/io/io_xdmf.py
@@ -1,0 +1,181 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of PyHOPE
+#
+# Copyright (c) 2024 Numerics Research Group, University of Stuttgart, Prof. Andrea Beck
+#
+# PyHOPE is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# PyHOPE is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# PyHOPE. If not, see <http://www.gnu.org/licenses/>.
+
+# ==================================================================================================================================
+# Mesh generation library
+# ==================================================================================================================================
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Standard libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+from functools import cache
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Third-party libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import h5py
+import numpy as np
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local imports
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local definitions
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ==================================================================================================================================
+
+
+@cache
+def xdmfElems(elemType: int, nGeo: int) -> tuple[str, int]:
+    """ XDMF: Get the element type from the HOPR format
+    - Linear
+        - Polyvertex - a group of unconnected points
+        - Polyline - a group of line segments
+        - Polygon
+        - Triangle
+        - Quadrilateral
+        - Tetrahedron
+        - Pyramid
+        - Wedge
+        - Hexahedron
+    - Quadratic
+        - Edge_3 - Quadratic line with 3 nodes
+        - Tri_6
+        - Quad_8
+        - Tet_10
+        - Pyramid_13
+        - Wedge_15
+        - Hex_20
+    - Arbitrary
+        - 1 - POLYVERTEX
+        - 2 - POLYLINE
+        - 3 - POLYGON
+        - 4 - TRIANGLE
+        - 5 - QUADRILATERAL
+        - 6 - TETRAHEDRON
+        - 7 - PYRAMID
+        - 8 - WEDGE
+        - 9 - HEXAHEDRON
+        - 16 - POLYHEDRON
+        - 34 - EDGE_3
+        - 35 - QUADRILATERAL_9
+        - 36 - TRIANGLE_6
+        - 37 - QUADRILATERAL_8
+        - 38 - TETRAHEDRON_10
+        - 39 - PYRAMID_13
+        - 40 - WEDGE_15
[…]
