### 2025-08-22 #8769 @jwallwork23
Some initial feedback on the paper:

* General feedback
	* PyHOPE sounds like a great piece of software providing functionality for advanced mesh generation.
	* The paper is well written and mostly conforms to the standard JOSS paper format.
	* The word count is slightly high so it could be beneficial to trim the text down a little.
* Summary section
	* The summary reads well, although I'm not sure the mention of space-filling curves is suitable for non-expert readers so perhaps the last sentence could be dropped? Interested readers could find this out by reading the HOPR paper cited.
* Statement of need section
	* I read the statement of need as being for a modern package that's readable and extensible.
* State of the field section
	* Usually JOSS papers have a "State of the field" section that follows the "Statement of need section". Please could you restructure the paper to move the "Related software" section there and rename it as such?
	* Could you give some examples of the "more complex mesh generation" functionality that makes PyHOPE unique amongst the other tools mentioned?
* Features section
	* Please could you provide a link for some information on the INI format? It's not something I've come across before.
	* The term "sanity check" was recently listed as a non-inclusive software term https://www.phoronix.com/news/LF-AOUSD-AWSF-Inclusive-Naming. Could I recommend using "consistency check" instead?
	* Would it be possible to provide some evidence for the scaling and parallel processing capabilities? Perhaps a figure with a comparison against similar tools? Alternatively, is there a paper that includes such a comparison that you could cite?

I'll do the software review separately and open any issues as they arise.
