### 2025-08-22 hopr-framework/PyHOPE#3 @joewallwork
[JOSS review]: Community guidelines
Towards https://github.com/openjournals/joss-reviews/issues/8769.

I can't see a community guidelines section. The JOSS review checklist includes
> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Please could you add such a section (or point me to where the existing one is)?
