### 2025-08-22 hopr-framework/PyHOPE#4 @jwallwork23
[JOSS review]: Functionality documentation
Towards https://github.com/openjournals/joss-reviews/issues/8769.

The JOSS review checklist includes
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Would it be possible to provide documentation for PyHOPE beyond the tutorials and usage documentation in the README?
