### 2025-08-22 hopr-framework/PyHOPE#5 @joewallwork
[JOSS review]: Automated tests
Towards https://github.com/openjournals/joss-reviews/issues/8769.

The JOSS review checklist includes
>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I see that there is automated GitLab CI testing defined in https://github.com/hopr-framework/PyHOPE/blob/main/.gitlab-ci.yml. Would it be possible to expose this in a test script so that users can run the tests to confirm their installation is working as expected?
