### 2025-08-25 c43c09c
XDMF: Add toggle, fix output for quadratic elements
