### 2025-08-25 b78c4f5
Feature: Health check (PyHOPE, packages, dependencies)
