+++ pyhope/check/check_health.py
@@ -1,0 +1,310 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of PyHOPE
+#
+# Copyright (c) 2024 Numerics Research Group, University of Stuttgart, Prof. Andrea Beck
+#
+# PyHOPE is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# PyHOPE is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# PyHOPE. If not, see <http://www.gnu.org/licenses/>.
+
+# ==================================================================================================================================
+# Mesh generation library
+# ==================================================================================================================================
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Standard libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import json
+import urllib.request
+from packaging.requirements import Requirement
+from packaging.version import Version
+from typing import Optional, Union
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Third-party libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local imports
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local definitions
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ==================================================================================================================================
+
+
+def PyPIVersion(package: str, timeout: int = 10) -> Optional[Version]:
+    """ Query PyPI JSON API for the latest version
+
+        Returns:
+            Optional[str]: Version string
+    """
+    url = f"https://pypi.org/pypi/{package}/json"
+    try:
+        with urllib.request.urlopen(url, timeout=timeout) as resp:
+            if resp.status != 200:
+                return None
+            data = json.load(resp)
+            info = data.get('info', {})
+            return Version(info.get('version'))
+    except Exception:
+        return None
+
+
+def _ParseVersion(text: str) -> Optional[Version]:
+    """ Find a semver-like version in arbitrary text
+    """
+    # Local imports ----------------------------------------
+    from re import search
+    from packaging.version import Version
+    # ------------------------------------------------------
+    if not text:
+        return None
+
+    # 1) Prefer to capture numeric version before a "-git" marker
+    #    Example: "4.14.0-git-8425b99f0" -> capture "4.14.0"
+    m = search(r"\bv?(\d+(?:\.\d+){0,3})(?=-git\b)", text)
+    if m:
+        try:
+            return Version(m.group(1))
+        except Exception:
+            return None
+
+    # 2) Fallback: capture semver-like token including -rc / +build etc.
+    #    Example: "v1.2.3-rc1", "1.2.3+build.1"
+    m = search(r"\bv?(\d+(?:\.\d+){0,3}(?:[-_a-zA-Z0-9+.]+)?)\b", text)
+    if m:
+        try:
[…]
