### 2025-08-25 80adba8
Bugfix: InnerBC are counted twice
