### 2025-08-25 e51e6ea
Improve wording for external Gmsh warning
