+++ pyhope/check/check_health.py
@@ -299,7 +299,7 @@ def CheckHealth() -> None:
 
     # For Gmsh, also check the packager
     gmshv, gmshp = GmshVersion()
-    if gmshp == 'NRG':
+    if gmshp.strip() == 'NRG':
         gmshi = ' (packaged by NRG)'
         gmshs = symbols.OK
     else:
@@ -307,4 +307,10 @@ def CheckHealth() -> None:
         gmshs = symbols.WARN
 
     DependencyHealth('Gmsh'    , version=gmshv,           status=gmshs, info=gmshi)
+    # Warn if we know that Gmsh uses an outdated CGNS
+    if gmshp.strip() != 'NRG':
+        print(hopout.warn('Detected Gmsh package uses an outdated CGNS (v3.4). ' +
+                          'For compatibility, replace with the updated NRG version',
+                          prefix=f'│ - {hopout.Symbols.WARN} ',
+                          warnonce=True))
     DependencyHealth('ParaView', version=DependencyVersion('paraview'), info=' (optional)')
