### 2025-08-26 hopr-framework/PyHOPE#4 @jwallwork23
> Would a documentation of the INI format and a more prominent link to the HOPR documentation be sufficient or did you have something else in mind?

I think that would be sufficient, thanks!
