### 2025-08-27 5a7bffd
bugfix run_in_parallel: make sure to propagate an exception from one pool process to the separate progress thread otherwise PyHOPE ends up in an idle state and cannot terminate properly (example: tutorial 2-02)
