### 2025-08-27 c7894a1
Feature: Installation check
[Download the tutorials], run them, and compare against saved TOML data
