+++ pyhope/check/check_install.py
@@ -1,0 +1,441 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of PyHOPE
+#
+# Copyright (c) 2024 Numerics Research Group, University of Stuttgart, Prof. Andrea Beck
+#
+# PyHOPE is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# PyHOPE is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# PyHOPE. If not, see <http://www.gnu.org/licenses/>.
+
+# ==================================================================================================================================
+# Mesh generation library
+# ==================================================================================================================================
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Standard libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import os
+import subprocess
+import sys
+import tempfile
+from typing import Final, Optional
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Third-party libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import h5py
+import numpy as np
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local imports
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local definitions
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ==================================================================================================================================
+
+
+def findGitRoot() -> Optional[str]:
+    """ Attempt to find the git root
+    """
+    try:
+        result = subprocess.run(['git', 'rev-parse', '--show-toplevel'],
+                                capture_output=True, text=True, check=True)
+        return result.stdout.strip()
+    except subprocess.CalledProcessError:
+        return None
+
+
+def downloadGitDir(user    : str,
+                   repo    : str,
+                   path    : str,
+                   target  : str,
+                   token   : Optional[str]  = None,
+                   branch  : str            = 'main',
+                   progress: Optional[bool] = True) -> None:
+    # Standard libraries -----------------------------------
+    import json
+    import time
+    import urllib.request
+    from urllib.error import HTTPError
+    # Local imports ----------------------------------------
+    import pyhope.output.output as hopout
+    from pyhope.common.common_progress import ProgressBar
+    # ------------------------------------------------------
+
+    # Helper to manage API requests and rate limiting
+    def _make_request(url : str,
+                      base: Optional[str]         = None,
+                      bar : Optional[ProgressBar] = None):
+
+        headers = {}
+        if token:
+            headers['Authorization'] = f'token {token}'
+
+        req = urllib.request.Request(url, headers=headers)
+
+        while True:
+            try:
+                return urllib.request.urlopen(req)
+            except HTTPError as e:
+                # Check for rate-limiting error
+                if  e.code == 403                        \
[…]
