+++ tutorials/3-02-sec_order_hex_CGNS_NACA/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 508032.0
+mean = 103522.83333333333
+stddev = 139199.81318943005
+
+[GlobalNodeIDs]
+min = 1.0
+max = 172116.0
+mean = 86016.70974466176
+stddev = 49518.71664624864
+
+[NodeCoords]
+min = -7304.063084226843
+max = 10000.000000000004
+mean = 750.9381098033388
+stddev = 2410.3581597396533
+
+[SideInfo]
+min = -57063.0
+max = 57064.0
+mean = 1931.9183071145126
+stddev = 15384.095723814895
