+++ tutorials/1-09-cartbox_nonconforming/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 14336.0
+mean = 4320.071428571428
+stddev = 4293.482273931573
+
+[GlobalNodeIDs]
+min = 1.0
+max = 2313.0
+mean = 1165.9151785714287
+stddev = 620.2568984198952
+
+[NodeCoords]
+min = -1.0
+max = 1.0
+mean = 0.07142857142882021
+stddev = 0.5767976926368594
+
+[SideInfo]
+min = -5632.0
+max = 5632.0
+mean = 182.51510416666667
+stddev = 1520.6530089007026
