+++ tutorials/3-03-sec_order_mixed_CGNS/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 17280.0
+mean = 3574.5714285714284
+stddev = 4693.7020402868775
+
+[GlobalNodeIDs]
+min = 1.0
+max = 6174.0
+mean = 3067.1666666666665
+stddev = 1709.5347817868137
+
+[NodeCoords]
+min = -10.0
+max = 20.0
+mean = 3.2452728597552607
+stddev = 7.914524093003563
+
+[SideInfo]
+min = -2060.0
+max = 2060.0
+mean = 81.8528855721393
+stddev = 554.9942909474032
