+++ tutorials/3-05-sec_order_layers_CGNS/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 58752.0
+mean = 13568.131193693695
+stddev = 15959.87359720015
+
+[GlobalNodeIDs]
+min = 1.0
+max = 10139.0
+mean = 5196.664232706971
+stddev = 3046.0353335423233
+
+[NodeCoords]
+min = -10.0
+max = 10.0
+mean = 1.4205013848782315
+stddev = 3.990519864176836
+
+[SideInfo]
+min = -11338.0
+max = 11340.0
+mean = 556.2442989634479
+stddev = 3179.3423719129723
