### 2025-08-27 01e9cfe
Add install check to cmdline option and runtime logic
