### 2025-08-27 997306a
Skip unsafe multiprocessing when running health checks
