+++ pyhope/script/pyhope_cli.py
@@ -99,6 +99,11 @@ def main() -> None:
     or args.verify_install:
         Check(args)
         sys.exit(0)
+
+    # Checks might be unsafe for multiprocessing
+    if args.skip_checks:
+        common._safeMTP = False
+
     # Check if there are unrecognized arguments
     if len(argv) >= 1:
         print('{} expects exactly one parameter or HDF5-mesh file! Exiting ...'
@@ -124,7 +129,8 @@ def main() -> None:
     hopout.sep()
 
     EliminateDuplicates()
-    OrientMesh()
+    if not args.skip_checks:
+        OrientMesh()
 
     # Build our data structures
     GenerateSides()
@@ -136,9 +142,10 @@ def main() -> None:
     FEMConnect()
 
     # Perform the mesh checks
-    CheckConnect()
-    CheckWatertight()
-    CheckJacobians()
+    if not args.skip_checks:
+        CheckConnect()
+        CheckWatertight()
+        CheckJacobians()
 
     # Output the mesh
     IO()
