### 2025-08-27 72757c3
CI/CD: Adjust TOML files for rounding issues
