+++ tutorials/1-01-cartbox/analyze.toml
@@ -1,23 +1,23 @@
 [ElemInfo]
 min = 0.0
-max = 30720.0
-mean = 7202.166666666667
-stddev = 8352.876984408826
+max = 512000.0
+mean = 85880.16666666667
+stddev = 147488.37048664398
 
 [GlobalNodeIDs]
 min = 1.0
-max = 4913.0
-mean = 2456.9656901041667
-stddev = 1348.1451518476915
+max = 389017.0
+mean = 194509.22983398437
+stddev = 110963.76583530543
 
 [NodeCoords]
 min = 0.0
 max = 1.0
-mean = 0.4999999999998388
-stddev = 0.2909216675784165
+mean = 0.49999999999983874
+stddev = 0.2891758724117992
 
 [SideInfo]
-min = -6525.0
-max = 6528.0
-mean = 335.3270670572917
-stddev = 1819.6277803871585
+min = -1725.0
+max = 1728.0
+mean = 74.57278645833334
+stddev = 455.5170187134046
