### 2025-08-27 277c7af
Revert: Skip unsafe multiprocessing when running health checks
