### 2025-08-27 3b365de
Update documentation with INI, health checks
