### 2025-08-27 44c857f
Cmdline: Actually exempt hidden arguments from help output
