+++ tutorials/1-08-curved-sphere/parameter.ini
@@ -1,0 +1,129 @@
+DEFVAR=(INT):    i0 = 004     ! no. elems in inner cube i0xi0xi0
+DEFVAR=(INT):    ir = 003     ! no. elems in r inside
+DEFVAR=(INT):    jr = 006     ! no. elems in r outside
+DEFVAR=(REAL):   ri = 0.5     ! inner  square dim
+DEFVAR=(REAL):   rm = 1.      ! middle square dim
+DEFVAR=(REAL):   r0 = 4.      ! outer  square dim
+DEFVAR=(REAL):   f1 = 1.2     ! strech factor first  ring
+DEFVAR=(REAL):   f2 = 1.2     ! strech factor second ring
+
+!================================================================================================================================= !
+! OUTPUT
+!================================================================================================================================= !
+ProjectName  = 1-08-sphere                          ! Name of output files
+Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
+
+!================================================================================================================================= !
+! MESH
+!================================================================================================================================= !
+Mode         = Internal                             ! Mode for Cartesian boxes
+NGeo         = 4                                    ! Polynomial order of the mapping
+nZones       = 13                                   ! Number of boxes
+
+! center
+Corner       = (/-ri,-ri,-ri ,,ri,-ri,-ri ,,ri,ri,-ri ,,  -ri,ri,-ri,, -ri,-ri,ri ,,ri,-ri,ri ,,ri,ri,ri ,,  -ri,ri,ri /)
+nElems       = (/i0,i0,i0/)                         ! Number of elements in each direction
+BCIndex      = (/0,0,0,0,0,0/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+Factor       = (/1.,1.,1./)                         ! Element stretching, either with a constant growth factor (+/- changes direction)
+! left (-x)
+Corner       = (/-rm,-rm,-rm ,,-ri,-ri,-ri ,,-ri,ri,-ri ,,  -rm,rm,-rm,, -rm,-rm,rm ,,-ri,-ri,ri ,,-ri,ri,ri ,,  -rm,rm,rm /)
+nElems       = (/ir,i0,i0/)                         ! Number of elements in each direction
+BCIndex      = (/0,0,0,0,0,0/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+Factor       = (/-f1,1.,1./)                        ! Element stretching, either with a constant growth factor (+/- changes direction)
+! right (+x)
+Corner       = (/ri,-ri,-ri ,,rm,-rm,-rm ,,rm,rm,-rm ,,  ri,ri,-ri,, ri,-ri,ri ,,rm,-rm,rm ,,rm,rm,rm ,,  ri,ri,ri /)
+nElems       = (/ir,i0,i0/)                         ! Number of elements in each direction
+BCIndex      = (/0,0,0,0,0,0/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+Factor       = (/f1,1.,1./)                         ! Element stretching, either with a constant growth factor (+/- changes direction)
+! lower (-y)
[…]
