+++ tutorials/1-08-curved-sphere/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 116000.0
+mean = 20297.330459770114
+stddev = 32963.348084757534
+
+[GlobalNodeIDs]
+min = 1.0
+max = 60281.0
+mean = 30134.372862068965
+stddev = 17293.308334402507
+
+[NodeCoords]
+min = -2.0
+max = 2.0
+mean = -0.002304199398700667
+stddev = 0.5604248254465676
+
+[SideInfo]
+min = -2830.0
+max = 2832.0
+mean = 104.39727011494253
+stddev = 761.4726810886073
