+++ tutorials/1-07-curved-torus/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 12288.0
+mean = 2280.527777777778
+stddev = 3426.4279625525073
+
+[GlobalNodeIDs]
+min = 1.0
+max = 5875.0
+mean = 2934.2105305989585
+stddev = 1662.6456197571297
+
+[NodeCoords]
+min = -6.0
+max = 6.0
+mean = 0.031206087055271378
+stddev = 2.908074017320844
+
+[SideInfo]
+min = -638.0
+max = 640.0
+mean = 30.497635135135134
+stddev = 169.8852533880788
