+++ tutorials/1-07-curved-torus/parameter.ini
@@ -1,0 +1,70 @@
+DEFVAR=(INT):    i0 = 002     ! no. elems in inner square i0xi0
+DEFVAR=(INT):    ir = 002     ! no. elems in r
+DEFVAR=(INT):    iz = 008     ! no. elems in toroidal dir
+DEFVAR=(REAL):   ri = 0.5     ! inner square  dim
+DEFVAR=(REAL):   rm = 1.      ! middle square dim
+DEFVAR=(REAL):   r0 = 2.      ! outer square  dim
+DEFVAR=(REAL):   s0 = 0.5     ! = 1/r0
+DEFVAR=(REAL):   rz = 5.      ! torus radius
+
+!================================================================================================================================= !
+! OUTPUT
+!================================================================================================================================= !
+ProjectName  = 1-07-torus                           ! Name of output files
+Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
+
+!================================================================================================================================= !
+! MESH
+!================================================================================================================================= !
+Mode         = Internal                             ! Mode for Cartesian boxes
+NGeo         = 3                                    ! Polynomial order of the mapping
+nZones       = 5                                    ! Number of boxes
+
+Corner       = (/-ri,-ri,0. ,,ri,-ri,0. ,,ri,ri,0. ,,  -ri,ri,0.,, -ri,-ri,1. ,,ri,-ri,1. ,,ri,ri,1. ,,  -ri,ri,1. /)
+nElems       = (/i0,i0,iz/)                         ! Number of elements in each direction
+BCIndex      = (/1,0,0,0,0,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+! left
+Corner       = (/-r0,-r0,0. ,,-ri,-ri,0. ,,-ri,ri,0. ,,  -r0,r0,0.,, -r0,-r0,1. ,,-ri,-ri,1. ,,-ri,ri,1. ,,  -r0,r0,1. /)
+nElems       = (/ir,i0,iz/)                         ! Number of elements in each direction
+BCIndex      = (/1,0,0,0,2,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+! right
+Corner       = (/ri,-ri,0. ,,r0,-r0,0. ,,r0,r0,0. ,,  ri,ri,0.,, ri,-ri,1. ,,r0,-r0,1. ,,r0,r0,1. ,,  ri,ri,1. /)
+nElems       = (/ir,i0,iz/)                         ! Number of elements in each direction
+BCIndex      = (/1,0,2,0,0,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+! upper
+Corner       = (/-ri,ri,0. ,,ri,ri,0. ,,r0,r0,0. ,,  -r0,r0,0.,, -ri,ri,1. ,,ri,ri,1. ,,r0,r0,1. ,,  -r0,r0,1. /)
+nElems       = (/4,ir,iz/)                          ! Number of elements in each direction
+BCIndex      = (/1,0,0,2,0,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
[…]
