### 2025-08-28 6aa1556
Postdeform: Cylinder and torus share the same logic
