+++ pyhope/mesh/transform/templates/cylinder.py
@@ -43,72 +43,90 @@ def PostDeform(points: np.ndarray) -> np.ndarray:
     This function applies a deformation transformation to the input points based on the given Fortran logic.
     The transformation maps a 2D square region to a cylindrical or toroidal coordinate system.
     """
+    # Local imports ----------------------------------------
+    import pyhope.output.output as hopout
+    from pyhope.readintools.readintools import CreateReal, GetReal
+    # ------------------------------------------------------
 
-    # TODO: Readin parameters from a configuration file
-    PostDeform_R0 = 1.0       # Define appropriate values
-    PostDeform_Rtorus = -1.0  # Adjust as needed
-    PostDeform_Lz = 1.0       # Cylinder height scale
-    PostDeform_sq = 0.0       # Spiral rotation parameter
-    MeshPostDeform = 1        # Deformation mode
+    hopout.sep()
+    CreateReal('PostDeform_R0',     default= 1.0, help='Scaling factor for the cylinder/torus radius')                         # noqa: E251
+    CreateReal('PostDeform_sq',     default= 0.0, help='Spiral factor along z. 0: no spiral, 1: one rotation for z in [0,1]')  # noqa: E251
+    CreateReal('PostDeform_Rtorus', default=-1.0, help='>0 for torus major radius (around z); <0 for cylinder mode')           # noqa: E251
+    CreateReal('PostDeform_Lz',     default= 1.0, help='Axial scaling for cylinder mode')                                      # noqa: E251
+
+    PostDeform_R0     = GetReal('PostDeform_R0')
+    PostDeform_sq     = GetReal('PostDeform_sq')
+    PostDeform_Rtorus = GetReal('PostDeform_Rtorus')
+    PostDeform_Lz     = GetReal('PostDeform_Lz')
 
     nTotal = points.shape[0]
-    X_out = np.zeros_like(points)
+    X_out  = np.zeros_like(points)
+    Pi     = np.pi
 
     for i in range(nTotal):
-        x = points[i, :].copy()
+        x  = points[i, :].copy()
         rr = max(abs(x[0]), abs(x[1]))
 
         if rr < 0.5:
-            dx1 = np.array([
-                0.5 * np.sqrt(2) * np.cos(0.25 * np.pi * x[1] / 0.5) - 0.5,
-                0.5 * np.sqrt(2) * np.sin(0.25 * np.pi * x[1] / 0.5) - x[1]
-            ])
-            dx2 = np.array([
-                0.5 * np.sqrt(2) * np.sin(0.25 * np.pi * x[0] / 0.5) - x[0],
-                0.5 * np.sqrt(2) * np.cos(0.25 * np.pi * x[0] / 0.5) - 0.5
-            ])
+            # inside [-0.5,0.5]^2
+            # right side at x=0.5
+            dx1_0 = 0.5 * np.sqrt(2.0) * np.cos(0.25 * Pi * x[1] / 0.5) - 0.5
+            dx1_1 = 0.5 * np.sqrt(2.0) * np.sin(0.25 * Pi * x[1] / 0.5) - x[1]
+            dx1 = np.array([dx1_0, dx1_1])
+
+            # upper side at y=0.5
+            dx2_0 = 0.5 * np.sqrt(2.0) * np.sin(0.25 * Pi * x[0] / 0.5) - x[0]
+            dx2_1 = 0.5 * np.sqrt(2.0) * np.cos(0.25 * Pi * x[0] / 0.5) - 0.5
+            dx2 = np.array([dx2_0, dx2_1])
+
[…]
