+++ tutorials/1-13-curved_periodic_hill/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 131072.0
+mean = 23928.166666666668
+stddev = 36784.532756838016
+
+[GlobalNodeIDs]
+min = 1.0
+max = 60025.0
+mean = 30014.710418701172
+stddev = 17001.52278169127
+
+[NodeCoords]
+min = 0.0
+max = 9.0
+mean = 2.7948472471761776
+stddev = 2.208208550209938
+
+[SideInfo]
+min = -6272.0
+max = 6272.0
+mean = 221.87063802083333
+stddev = 1687.11359440716
