+++ tutorials/1-08-curved_sphere/parameter.ini
@@ -107,7 +107,7 @@ BoundaryName = BC_zminus                            ! BC index 1 (from  position
 BoundaryType = (/2,0,0,0/)                          ! (/ Type, curveIndex, State, alpha /)
 BoundaryName = BC_yminus                            ! BC index 2
 BoundaryType = (/2,0,0,0/)
-BoundaryName = BC_xplus                             ! BC index 3 ...
+BoundaryName = BC_xplus                             ! BC index 3
 BoundaryType = (/2,0,0,0/)
 BoundaryName = BC_yplus                             ! BC index 4
 BoundaryType = (/2,0,0,0/)
