+++ tutorials/1-07-curved_torus/parameter.ini
@@ -53,7 +53,7 @@ BoundaryName = BC_zminus                            ! BC index 1 (from  position
 BoundaryType = (/1,0,0,1/)                          ! (/ Type, curveIndex, State, alpha /)
 BoundaryName = BC_outer                             ! BC index 2
 BoundaryType = (/2,1,0,0/)
-BoundaryName = BC_zplus                             ! BC index 6
+BoundaryName = BC_zplus                             ! BC index 3
 BoundaryType = (/1,0,0,-1/)
 vv = (/0.,0.,1./)                                   ! Vector for periodic BC
 !================================================================================================================================= !
