+++ tutorials/1-13-curved_periodic_hill/parameter.ini
@@ -1,0 +1,51 @@
+
+! Parameters for the Ercoftac periodic hill testcase
+! http://www.kbwiki.ercoftac.org/w/index.php/Abstr:2D_Periodic_Hill_Flow
+! For Re=700, a resolution of 16x16x8 elems with bell stretching and DXmaxToDXmin = (/1.5,10.,1./)
+! yields approximately y+=0.8 and x+=z+=12 at the hill cusp and about half of that everywhere else.
+
+!=============================================================================== !
+! OUTPUT
+!=============================================================================== !
+ProjectName  = 1-13-curved_periodic_hill            ! Name of output files
+Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
+
+!=============================================================================== !
+! MESH
+!=============================================================================== !
+Mode         = Internal                             ! Mode for Cartesian boxes
+NGeo         = 3                                    ! Polynomial order of the mapping
+nZones       = 1                                    ! Number of boxes
+
+X0           = (/0.,0.   ,0. /)                     ! Origin of a zone. Equivalent to a corner node.
+DX           = (/9.,3.035,4.5/)                     ! Extension of the zone in each spatial direction starting from the origin X0 corner node
+nElems       = (/16,16   ,8  /)                     ! Number of elements in each direction
+BCIndex      = (/1,2,3,4,5,6/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (108: Hexahedral)
+
+StretchType  = (/3  ,3  ,0 /)                       ! Stretching type for individual zone per spatial direction.
+DXmaxToDXmin = (/1.5,10.,1./)                       ! Ratio between the smallest and largest element per spatial direction
+
+!================================================================================================================================= !
+! MESH POST DEFORM
+!================================================================================================================================= !
+MeshPostDeform    = PHill                           ! Periodic hill Ercoftac test case
+
+!=============================================================================== !
+! BOUNDARY CONDITIONS
+!=============================================================================== !
+BoundaryName = BC_periodicz-                        ! BC index 1 (from  position in parameterfile)
+BoundaryType = (/1,0,0,2/)                          ! (/ Type, curveIndex, State, alpha /)        ! (/bc-type,curved index,bc state,displacement vector/)
+BoundaryName = BC_wall_lower                        ! BC index 2
+BoundaryType = (/4,0,1,0/)
+BoundaryName = BC_periodicx+                        ! BC index 3
+BoundaryType = (/1,0,0,-1/)
+BoundaryName = BC_wall_upper                        ! BC index 4
+BoundaryType = (/4,0,1,0/)
+BoundaryName = BC_periodicx-                        ! BC index 5
+BoundaryType = (/1,0,0,1/)
+BoundaryName = BC_periodicz+                        ! BC index 6
+BoundaryType = (/1,0,0,-2/)                         ! (/bc-type,curved index,bc state,displacement vector/)
+vv = (/9.,0.,0. /)                                  ! Vector for periodic BC
+vv = (/0.,0.,4.5/)                                  ! Vector for periodic BC
