### 2025-08-28 6218240
Add 1-06 tutorial
