+++ tutorials/1-06-curved_cylinder/parameter.ini
@@ -1,16 +1,15 @@
 DEFVAR=(INT):    i0 = 002     ! no. elems in inner square i0xi0
 DEFVAR=(INT):    ir = 002     ! no. elems in r
-DEFVAR=(INT):    iz = 008     ! no. elems in toroidal dir
+DEFVAR=(INT):    iz = 004     ! no. elems in z
 DEFVAR=(REAL):   ri = 0.5     ! inner square  dim
 DEFVAR=(REAL):   rm = 1.      ! middle square dim
 DEFVAR=(REAL):   r0 = 2.      ! outer square  dim
-DEFVAR=(REAL):   s0 = 0.5     ! = 1/r0
-DEFVAR=(REAL):   rz = 5.      ! torus radius
+DEFVAR=(REAL):   lz = 20.     ! length of domain in z
 
 !================================================================================================================================= !
 ! OUTPUT
 !================================================================================================================================= !
-ProjectName  = 1-07-torus                           ! Name of output files
+ProjectName  = 1-06-curved_cylinder                 ! Name of output files
 Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 
@@ -18,32 +17,33 @@ OutputFormat = HDF5                                 ! Mesh output format (HDF5 V
 ! MESH
 !================================================================================================================================= !
 Mode         = Internal                             ! Mode for Cartesian boxes
-NGeo         = 3                                    ! Polynomial order of the mapping
+NGeo         = 4                                    ! Polynomial order of the mapping
 nZones       = 5                                    ! Number of boxes
 
-Corner       = (/-ri,-ri,0. ,,ri,-ri,0. ,,ri,ri,0. ,,  -ri,ri,0.,, -ri,-ri,1. ,,ri,-ri,1. ,,ri,ri,1. ,,  -ri,ri,1. /)
+! center
+Corner       = (/-ri,-ri,0. ,,ri,-ri,0. ,,ri,ri,0. ,,  -ri,ri,0.,, -ri,-ri,lz ,,ri,-ri,lz ,,ri,ri,lz ,,  -ri,ri,lz /)
 nElems       = (/i0,i0,iz/)                         ! Number of elements in each direction
-BCIndex      = (/1,0,0,0,0,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+BCIndex      = (/1,0,0,0,0,6/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
 ElemType     = 108                                  ! Element type (108: Hexahedral)
 ! left
-Corner       = (/-r0,-r0,0. ,,-ri,-ri,0. ,,-ri,ri,0. ,,  -r0,r0,0.,, -r0,-r0,1. ,,-ri,-ri,1. ,,-ri,ri,1. ,,  -r0,r0,1. /)
+Corner       = (/-r0,-r0,0. ,,-ri,-ri,0. ,,-ri,ri,0. ,,  -r0,r0,0.,, -r0,-r0,lz ,,-ri,-ri,lz ,,-ri,ri,lz ,,  -r0,r0,lz /)
 nElems       = (/ir,i0,iz/)                         ! Number of elements in each direction
-BCIndex      = (/1,0,0,0,2,3/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
[…]
