### 2025-08-29 64f58d6
Documentation: Replace CONTRIBUTORS.md with AUTHORS.md
