+++ AUTHORS.md
@@ -1,0 +1,21 @@
+# Authors
+
+PyHOPE's development is coordinated by Patrick Kopper, who is the *principal developer*
+and main contributor. In addition, there are *contributors* who have
+provided substantial additions or modifications. Together, these two groups form
+authors as mentioned in the [LICENSE.md](LICENSE.md) file.
+
+## Principal Developers
+* [Patrick Kopper](https://www.iag.uni-stuttgart.de/en/institute/team/Kopper-00001/),
+  University of Stuttgart, Germany
+
+## Contributors
+The following people contributed major additions or modifications to PyHOPE and
+are listed in alphabetical order:
+
+* Marcel Blind
+* Stephen Copplestone
+* Patrick Kopper
+* Marius Kurz
+* Felix Rodach
+* Anna Schwarz
