### 2025-08-29 e7b5292
Documentation: Add CONTRIBUTING.md
