### 2025-08-29 hopr-framework/PyHOPE#5 @kopperp
Thanks for the praise and the catch! Really appreciate your feedback and you taking the time to create corresponding issues.

I am still working on the community guidelines, I will include the fix there.
