### 2025-08-29 5f3dd30
Documentation: Update readme with verification, documentation, and licence sections
