+++ README.md
@@ -26,31 +26,32 @@ PyHOPE is built using standard Python packages. You can install PyHOPE by follow
     python -m pip install pyhope
     ```
 
-3.  **Optional: Verify PyHOPE installation**  
-    PyHOPE features internal health checks. The checks can be invoked directly from the terminal.
-    ```bash
-    pyhope --verify [tutorials]           # Run all health checks
-    pyhope --verify-health                # Run Python health checks
-    pyhope --verify-install [tutorialss]  # Run PyHOPE mesh generation checks
-    ```
+# Testing
+PyHOPE features internal health checks to verify that everything works as expected. The checks can be invoked directly from the terminal.
+```bash
+pyhope --verify [tutorials]          # Run all health checks
+pyhope --verify-health               # Run Python health checks
+pyhope --verify-install [tutorials]  # Run PyHOPE mesh generation checks
+```
 
 > [!NOTE]  
 >  By default, PyHOPE looks for the `tutorials` directory relative to the current working directory or the git root. If neither exists, PyHOPE downloads the tests from GitHub while using available authentication methods.
 
-4. **Run PyHOPE**  
-    PyHOPE is available as a command-line tool. After installation, its functionalities can be accessed directly from the terminal by passing a valid configuration file.
-    ```bash
-    pyhope [parameter.ini]
-    ```
+# Getting Help
+PyHOPE help output is formatted to serve as self-hosting INI format. A list of all options and the default values can be accessed by running the following command.
+```bash
+pyhope --help
+```
 
-> [!TIP]
-> PyHOPE help output is formatted to serve as self-hosting INI format. A list of all options and the default values can be accessed by running the following command.
-> ```bash
-> pyhope --help
-> ```
+# Documentation
+PyHOPE is heavily inspired by [HOPR (High Order Preprocessor)](https://github.com/hopr-framework/hopr) and shares the same input/output format. For more information and tutorials, please visit the [HOPR documentation](https://hopr.readthedocs.io).
 
 # Usage
-PyHOPE is invoked from the command line. Run parameters are read from a configuration file. The following output is obtained when running the example configuration file `tutorials/1-01-cartbox/parameter.ini`.
+PyHOPE is invoked from the command line. After installation, its functionalities can be accessed directly from the terminal by passing a valid configuration file.
+```bash
+pyhope [parameter.ini]
+```
+The following output is obtained when running the example configuration file `tutorials/1-01-cartbox/parameter.ini`.
 ```
 $ pyhope tutorials/1-01-cartbox/parameter.ini
 ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -167,3 +168,6 @@ $ pyhope tutorials/1-01-cartbox/parameter.ini
 ┃ PyHOPE completed in [0.25 sec]
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 ```
+
+# Licence
+PyHOPE is licensed under the GPL-3.0 license (see [LICENSE.md](LICENSE.md)).
