### 2025-08-29 e175bf9
Documentation: Update readme with command line/context manager usage
