### 2025-08-29 3e8d79b
Packaging: Remove unused sortedcontainers
