### 2025-08-29 83e14e3
CI/CD: Add code coverage to output
