### 2025-08-29 0a1b748
CI/CD: Update bookworm to trixie
