+++ .gitlab-ci.yml
@@ -24,8 +24,8 @@
 
 # Official language image. Look for the different tagged releases at:
 # > https://hub.docker.com/r/library/python/tags/
-# image: ${CI_REGISTRY_NAME}/nrg-python:3.13.2-bookworm-x86_64
-image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.13.2-bookworm-x86_64
+# image: ${CI_REGISTRY_NAME}/nrg-python:3.13.7-trixie-x86_64
+image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.13.7-trixie-x86_64
 
 # Change pip's cache directory to be inside the project directory since we can only cache local items.
 variables:
@@ -128,25 +128,25 @@ ruff:
 python:3.10:
   extends: .defaults
   stage: compatibility
-  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.10.16-bookworm-x86_64
+  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.10.18-trixie-x86_64
   <<: *compatibility-script
 # Python 3.11 ----------------------------------------------------------------------------------------------------------------------------------------
 python:3.11:
   extends: .defaults
   stage: compatibility
-  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.11.11-bookworm-x86_64
+  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.11.13-trixie-x86_64
   <<: *compatibility-script
 # Python 3.12 ----------------------------------------------------------------------------------------------------------------------------------------
 python:3.12:
   extends: .defaults
   stage: compatibility
-  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.12.9-bookworm-x86_64
+  image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.12.11-trixie-x86_64
   <<: *compatibility-script
 # Python 3.13 ----------------------------------------------------------------------------------------------------------------------------------------
 python:3.13:
   extends: .defaults
   stage: compatibility
-  # image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.13.2-bookworm-x86_64 # default version at the moment
+  # image: registry.iag.uni-stuttgart.de/flexi/codes/pyhope/nrg-python:3.13.7-trixie-x86_64 # default version at the moment
   <<: *compatibility-script
 # # Python 3.14 ----------------------------------------------------------------------------------------------------------------------------------------
 # python_314:
