+++ .docker/python_3.12.11-trixie/dockerfile
@@ -1,5 +1,5 @@
 # Load the base image
-FROM python:3.12.9-bookworm
+FROM python:3.12.11-trixie
 
 # Numerics research group
 LABEL maintainer="numerics@iag.uni-stuttgart.de"
@@ -31,6 +31,6 @@ RUN apt-get update              && \
 # Setup uv package manager
 RUN mkdir -p /.local/bin        && \
     curl -fsSL -o /tmp/uv.tar.gz   \
-      https://github.com/astral-sh/uv/releases/download/0.7.15/uv-x86_64-unknown-linux-gnu.tar.gz && \
+      https://github.com/astral-sh/uv/releases/download/0.8.14/uv-x86_64-unknown-linux-gnu.tar.gz && \
     tar -xzf /tmp/uv.tar.gz -C /.local/bin --strip-components=1 && \
     rm /tmp/uv.tar.gz
