### 2025-08-31 9de31a8
Feature: Debug mesh output
Add optional VTK visualization of the volume and surfaces meshes, including derivated quantities
