+++ pyhope/io/io_vars.py
@@ -41,6 +41,7 @@ projectname  : str                               # Name of output files
 outputformat : int                               # Mesh output format
 outputmeta   : bool                              # Mesh output metadata (optional)
 
+debugmesh    : bool                              # Mesh output debug mesh
 debugvisu    : bool                              # Enable and show debug output / visualization
 
 
