+++ pyhope/io/io_debug.py
@@ -1,0 +1,249 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of PyHOPE
+#
+# Copyright (c) 2024 Numerics Research Group, University of Stuttgart, Prof. Andrea Beck
+#
+# PyHOPE is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# PyHOPE is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# PyHOPE. If not, see <http://www.gnu.org/licenses/>.
+
+# ==================================================================================================================================
+# Mesh generation library
+# ==================================================================================================================================
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Standard libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+from typing import Final, Union
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Third-party libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import meshio
+import numpy as np
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local imports
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local definitions
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ==================================================================================================================================
+
+
+def writeVTM(filename: str,
+             blocks  : Union[list, tuple]) -> None:
+    # Standard libraries -----------------------------------
+    import xml.etree.ElementTree as ET
+    # ------------------------------------------------------
+    # blocks is a list of tuples: (index, name, filepath)
+    vtkfile = ET.Element('VTKFile', attrib={'type'      : 'vtkMultiBlockDataSet',
+                                            'version'   : '1.1',
+                                            'byte_order': 'LittleEndian'})
+
+    multiblock = ET.SubElement(vtkfile, 'vtkMultiBlockDataSet')
+
+    for idx, name, filepath in blocks:
+        ET.SubElement(multiblock, 'DataSet', attrib={'index': str(idx),
+                                                     'name' : name,
+                                                     'file' : filepath})
+
+    tree = ET.ElementTree(vtkfile)
+    tree.write(filename, encoding='utf-8', xml_declaration=True)
+
+
+def DebugIO() -> None:
+    """ Routine to output the debug mesh. Downcast the existing
+        PyHOPE format to first order, enrich with debug information
+        and output
+    """
+    # Local imports ----------------------------------------
+    import pyhope.io.io_vars as io_vars
[…]
