+++ README.md
@@ -63,6 +63,8 @@ $ pyhope tutorials/1-01-cartbox/parameter.ini
 │ INIT OUTPUT...
 │                     ProjectName │ 1-01-cartbox                    │ *CUSTOM │
 │                    OutputFormat │ 0 [HDF5]                        │ *CUSTOM │
+│                  OutputMetadata │ True                            │ DEFAULT │
+│                       DebugMesh │ T                               │ *CUSTOM │
 │                       DebugVisu │ F                               │ *CUSTOM │
 ├─────────────────────────────────────────────
 │ INIT MESH...
@@ -163,6 +165,10 @@ $ pyhope tutorials/1-01-cartbox/parameter.ini
 │         Curved Hexahedra  :          512
 ├────
 ├── Writing HDF5 mesh to "1-01-cartbox_mesh.h5"
+│  WARNING  ┃ XDMF only supports linear/quadratic elements, skipping...
+├────
+├── Writing volume  debug mesh to "1-01-cartbox_DebugElem.vtu"
+├── Writing surface debug mesh to "1-01-cartbox_DebugSide.vtu"
 ┢━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 ┃ PyHOPE completed in [0.25 sec]
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
