+++ tutorials/2-03-external_mesh_Gmsh_curved/parameter.ini
@@ -2,7 +2,7 @@
 ! OUTPUT
 !=============================================================================== !
 ProjectName  = 2-03-external_mesh_Gmsh_curved       ! Name of output files
-Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
