+++ tutorials/5-03-convtest_mortar/parameter.ini
@@ -2,7 +2,7 @@
 ! OUTPUT
 !================================================================================================================================= !
 ProjectName  = 5-03-convtest_mortar                 ! Name of output files
-Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
