+++ tutorials/2-04-external_mesh_CGNS_tetra_splitToHex/parameter.ini
@@ -2,7 +2,7 @@
 ! OUTPUT
 !=============================================================================== !
 ProjectName  = 2-04-external_mesh_CGNS_tetra        ! Name of output files
-Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
