+++ tutorials/5-01-convtest/parameter.ini
@@ -2,7 +2,7 @@
 ! OUTPUT
 !================================================================================================================================= !
 ProjectName  = 5-01-convtest                        ! Name of output files
-Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 

[…]
