### 2025-09-01 20f3da9
CI/CD: Improve health check formatting
