+++ docs/index.md
@@ -1,0 +1,17 @@
+# PyHOPE
+
+Mesh Generator for High-Order Meshes in Python
+
+---
+
+PyHOPE (Python High-Order Preprocessing Environment) is an open-source Python framework for the generation of three-dimensional unstructured high-order meshes. These meshes are needed by high-order numerical methods like Discontinuous Galerkin, Spectral Element Methods, or pFEM, in order to retain their accuracy if the computational domain includes curved boundaries.
+
+<div class="text-center" style="margin-bottom: 2em;">
+<a href="getting-started/" class="btn btn-primary" role="button">Getting Started</a>
+<a href="user-guide/"      class="btn btn-primary" role="button">User Guide</a>
+<a href="mesh-format/"     class="btn btn-primary" role="button">Mesh Format</a>
+</div>
+
+PyHOPE has been developed by the Numerics Research Group (NRG) lead by Prof. Andrea Beck at the Institute of Aerodynamics and Gas Dynamics at the University of Stuttgart, Germany. 
+
+This is a scientific project. If you use PyHOPE for publications or presentations in science, please support the project by citing our publications given at [numericsresearchgroup.org](https://numericsresearchgroup.org/publications.html).
