### 2025-09-01 6282cae
Documentation: Add MkDocs to Gitlab CI/CD
