### 2025-09-01 fed4ecb
Documentation: Add MkDocs to GitHub Actions, deploy to GitHub Pages
