### 2025-09-01 93b2a99
Documentation: Add installation backlink
