### 2025-09-01 ca71e48
Coverage: Exclusion rules
