### 2025-09-01 72a03be
adopt GetPiMultiplies feature from HOPR, as needed by e.g. TGV tutorial in FLEXI: parse coordinate 'fPi' as f*PI with f being a float
