### 2025-09-02 4cc92e5
rename GetPiMultiplies to ParseFloatAndPi, added tutorial for this function
