+++ tutorials/1-05_cartbox_pimultiples/parameter.ini
@@ -1,0 +1,39 @@
+!================================================================================================================================= !
+! OUTPUT
+!================================================================================================================================= !
+ProjectName  = 1-05-pimultiples                     ! Name of output files
+Debugvisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
+! nThreads     = 4                                    ! Number of threads for multiprocessing
+
+!================================================================================================================================= !
+! MESH
+!================================================================================================================================= !
+Mode         = 1                                    ! Mode for Cartesian boxes
+nZones       = 1                                    ! Number of boxes
+Corner       = (/0.,0.,0. ,,Pi,0.,0. ,,Pi,2Pi,0. ,,  0.,2Pi,0.,, 0.,0.,1 ,,Pi,0.,1. ,,Pi,2Pi,1. ,,  0.,2Pi,1. /)
+                                                    ! Corner node positions: (/ x_1,y_1,z_1, x_2,y_2,z_2,..... , x_8,y_8,z_8/)
+nElems       = (/8,8,8/)                            ! Number of elements in each direction
+BCIndex      = (/1,2,3,4,5,6/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
+ElemType     = 108                                  ! Element type (104: Tetrahedra, 105: Pyramid, 106: Prism, 108: Hexahedral)
+nGeo         = 1
+MeshScale    = 3Pi
+
+!================================================================================================================================= !
+! BOUNDARY CONDITIONS
+!================================================================================================================================= !
+BoundaryName = BC_zminus                            ! BC index 1 (from position in parameterfile)
+BoundaryType = (/4,0,0,0/)                          ! (/ Type, curveIndex, State, alpha /)
+BoundaryName = BC_yminus                            ! BC index 2
+BoundaryType = (/2,0,0,0/)
+BoundaryName = BC_xplus                             ! BC index 3
+BoundaryType = (/2,0,0,0/)
+BoundaryName = BC_yplus                             ! BC index 4
+BoundaryType = (/2,0,0,0/)
+BoundaryName = BC_xminus                            ! BC index 5
+BoundaryType = (/2,0,0,0/)
+BoundaryName = BC_zplus                             ! BC index 6
+BoundaryType = (/9,0,0,0/)
+vv = (/Pi, 0., 0./)
+vv = (/0., 2Pi, 0./)
+vv = (/0., 0., 1./)
