+++ tutorials/1-05_cartbox_pimultiples/analyze.toml
@@ -1,0 +1,23 @@
+[ElemInfo]
+min = 0.0
+max = 4096.0
+mean = 1212.8333333333333
+stddev = 1201.5390708956945
+
+[GlobalNodeIDs]
+min = 1.0
+max = 729.0
+mean = 365.50830078125
+stddev = 191.83670202439635
+
+[NodeCoords]
+min = 0.0
+max = 59.21762640653615
+mean = 16.375202928427004
+stddev = 15.25286653474704
+
+[SideInfo]
+min = -1725.0
+max = 1728.0
+mean = 74.57278645833334
+stddev = 455.5170187134045
