### 2025-09-02 f6614c3
CI/CD: Adjust TOML files for rounding issues
