### 2025-09-02 fb22ff9
Documentation: Add Daniel Appel
