### 2025-09-02 b50848c
include float as pi multiplier in tutorial 1-05 and update analyze.toml accordingly
