+++ tutorials/1-05_cartbox_pimultiples/analyze.toml
@@ -7,14 +7,14 @@ stddev = 1201.5390708956945
 [GlobalNodeIDs]
 min = 1.0
 max = 729.0
-mean = 365.50830078125
-stddev = 191.83670202439635
+mean = 365.0625
+stddev = 191.17161383819513
 
 [NodeCoords]
 min = 0.0
 max = 59.21762640653615
-mean = 16.375202928427004
-stddev = 15.25286653474704
+mean = 20.874213308296518
+stddev = 14.270222849305025
 
 [SideInfo]
 min = -1725.0
