+++ tutorials/1-05_cartbox_pimultiples/parameter.ini
@@ -11,7 +11,7 @@ OutputFormat = HDF5                                 ! Mesh output format (HDF5 V
 !================================================================================================================================= !
 Mode         = 1                                    ! Mode for Cartesian boxes
 nZones       = 1                                    ! Number of boxes
-Corner       = (/0.,0.,0. ,,Pi,0.,0. ,,Pi,2Pi,0. ,,  0.,2Pi,0.,, 0.,0.,1 ,,Pi,0.,1. ,,Pi,2Pi,1. ,,  0.,2Pi,1. /)
+Corner       = (/0.,0.,0. ,,Pi,0.,0. ,,Pi,2Pi,0. ,,  0.,2Pi,0.,, 0.,0.,1.23Pi ,,Pi,0.,1.23Pi ,,Pi,2Pi,1.23Pi ,,  0.,2Pi,1.23Pi /)
                                                     ! Corner node positions: (/ x_1,y_1,z_1, x_2,y_2,z_2,..... , x_8,y_8,z_8/)
 nElems       = (/8,8,8/)                            ! Number of elements in each direction
 BCIndex      = (/1,2,3,4,5,6/)                      ! Indices of boundary conditions for six boundary faces (z-,y-,x+,y+,x-,z+)
@@ -36,4 +36,4 @@ BoundaryName = BC_zplus                             ! BC index 6
 BoundaryType = (/9,0,0,0/)
 vv = (/Pi, 0., 0./)
 vv = (/0., 2Pi, 0./)
-vv = (/0., 0., 1./)
+vv = (/0., 0., 1.23Pi/)
