### 2025-09-02 859a0be
CI/CD: Only consider one job for coverage
