### 2025-09-02 bd82d47
Revert "CI/CD: Add health checks to coverage"
This reverts commit 19811dc8f6c745cbdec8c2225528cad95e9d9f69.
