### 2025-09-02 30712e9
Documentation: Make reference to HOPR more prominent
