### 2025-09-02 a204c68
Documentation: Explicitly link installation page
