### 2025-09-02 d41e421
CI/CD: Add combined coverage
