### 2025-09-02 94f2214
Linter: Remove dead code with vulture
