+++ .gmsh/build_manylinux.py
@@ -123,7 +123,7 @@ arch = platform.machine()
 WORK_DIR     = '/io'
 BUILD_DIR    = os.path.join(WORK_DIR, 'build')
 INSTALL_DIR  = os.path.join(WORK_DIR, 'gmsh_install')
-Doxygen_DIR  = os.path.join(WORK_DIR, 'doxygen')
+# Doxygen_DIR  = os.path.join(WORK_DIR, 'doxygen')
 
 # HDF5
 HDF5_DIR     = os.path.join(WORK_DIR, 'hdf5')
