### 2025-09-03 4a02d05
Improvement: Add coverage script, exclude (disabled) multiprocessing during coverage
