+++ tools/create_coverage.py
@@ -1,0 +1,283 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This file is part of PyHOPE
+#
+# Copyright (c) 2024 Numerics Research Group, University of Stuttgart, Prof. Andrea Beck
+#
+# PyHOPE is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# PyHOPE is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# PyHOPE. If not, see <http://www.gnu.org/licenses/>.
+
+# ==================================================================================================================================
+# Stub generation helper
+# ==================================================================================================================================
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Standard libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import glob
+import os
+import shutil
+import subprocess
+import sys
+# Third-party libraries
+# ----------------------------------------------------------------------------------------------------------------------------------
+import argparse
+from typing import List, Optional, Mapping, Tuple
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local imports
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ----------------------------------------------------------------------------------------------------------------------------------
+# Local definitions
+# ----------------------------------------------------------------------------------------------------------------------------------
+# ==================================================================================================================================
+
+# ANSI colors
+GREEN = '\033[32m'
+RED   = '\033[31m'
+CYAN  = '\033[36m'
+RESET = '\033[0m'
+
+
+def printHeader(title: str, width: int) -> None:
+    ''' Print a box header for a section.
+    '''
+    print('')
+    print('┌─' + '─' * width + '─┐')
+    print('│ ' + f'{title:<{width}}' + ' │')
+    print('└─' + '─' * width + '─┘')
+
+
+def findCoverageFiles(base_dir   : str,
+                      pattern    : str,
+                      output_name: str) -> List[str]:
+    ''' Find coverage data files in base_dir that match pattern, excluding the combined output
+    '''
+    candidates    = sorted(f for f in glob.glob(os.path.join(base_dir, pattern)) if os.path.isfile(f))
+    combined_path = os.path.join(base_dir, output_name)
+    combined_abs  = os.path.abspath(combined_path)
+    input_datafiles = [f for f in candidates if os.path.abspath(f) != combined_abs]
+    return input_datafiles
+
+
+def findParameterFiles(base_dir  : str,
+                       param_name: str) -> List[str]:
+    ''' Recursively find all parameter.ini files under tutorials_root
+    '''
+    matches: List[str] = []
+    for base_dir, _dirs, files in os.walk(base_dir):
+        if param_name in files:
+            matches.append(os.path.join(base_dir, param_name))
+    matches.sort()
+    return matches
+
+
+def runCmd(cmd: List[str],
+           cwd: str,
+           env: Optional[Mapping[str, str]] = None) -> None:
+    ''' Run a command, streaming output to the console. Raise on failure.
+    '''
+    try:
+        subprocess.run(cmd, cwd=cwd, env=env, check=True)
+    except subprocess.CalledProcessError:
[…]
