+++ pyhope/common/common_tools.py
@@ -42,8 +42,7 @@ import numpy as np
 # ==================================================================================================================================
 
 
-# pragma: no cover
-def time_function(func, *args, **kwargs) -> float:
+def time_function(func, *args, **kwargs) -> float:  # pragma: no cover
     """ A helper function to measure the execution time of an arbitrary function.
 
     Parameters:
