### 2025-09-03 1b197c9
Perf: Improve performance of duplicates detection
