+++ pyhope/mesh/mesh_duplicates.py
@@ -25,14 +25,15 @@
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
-import copy
 import gc
-from typing import Final, cast
+from collections import defaultdict
+from typing import Dict, Final, Tuple, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 import numpy as np
 from scipy.spatial import KDTree
+from scipy.sparse.csgraph import connected_components
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Local imports
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -40,6 +41,48 @@ from scipy.spatial import KDTree
 # Local definitions
 # ----------------------------------------------------------------------------------------------------------------------------------
 # ==================================================================================================================================
+
+
+def _findPointsTol(points: np.ndarray, tol: float) -> np.ndarray:
+    """ Build an undirected connectivity graph for points within 'tol', then compute
+        the connected components and pick the minimum index in each component as the
+        representative
+    """
+
+    # Create a KDTree for the mesh points
+    tree = KDTree(points)
+
+    # Construct a sparse adjacency matrix where edges connect points within 'tol'
+    # > Use sparse_distance_matrix to avoid Python-level loops and return COO/CSR in C
+    #
+    # NOTE: This includes self-distances (diagonal); zero them below
+    adj = tree.sparse_distance_matrix(tree, tol, output_type='coo_matrix')
+    adj = adj.tocsr()
+
+    # Remove self-connections, enforce symmetry
+    adj.setdiag(0)
+    adj.eliminate_zeros()
+    # Make matrix symmetric (in case of any asymmetry)
+    adj = adj.maximum(adj.T)
+
+    # Ensure canonical CSR for faster graph ops
+    adj.sum_duplicates()
+    adj.sort_indices()
+
+    # If there are no edges (all points isolated w.r.t. tol), each point is its own component
+    nPoints = points.shape[0]
+    if adj.nnz == 0:
+        return np.arange(nPoints, dtype=int)
+
+    # Compute connected components (undirected)
+    components, labels = connected_components(adj, directed=False, return_labels=True)
+
+    # For each component label, choose the minimum original point index as representative
+    repLabel  = np.full(components, nPoints, dtype=int)
+    np.minimum.at(repLabel, labels, np.arange(nPoints, dtype=int))
+    repsPoint = repLabel[labels]
+
+    return repsPoint
 
 
 def EliminateDuplicates() -> None:
@@ -61,28 +104,27 @@ def EliminateDuplicates() -> None:
     cdict: Final[dict] = mesh.cells_dict
 
     # Find the mapping to the (N-1)-dim elements
-    csetMap = { key: tuple(i for i, cell in enumerate(cset) if cell is not None and cast(np.ndarray, cell).size > 0)
-                             for key, cset in csets.items()}
+    csetMap: Dict      = { key: tuple(i for i, cell in enumerate(cset) if cell is not None and cast(np.ndarray, cell).size > 0)
+                                        for key, cset in csets.items()}
 
[…]
