### 2025-09-04 9ab9440
Perf: Improve performance of mesh sorting
