+++ pyhope/mesh/mesh_sort.py
@@ -26,11 +26,12 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 import gc
-from typing import cast, final
+from typing import Final, cast, final
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 import numpy as np
+import numpy.typing as npt
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Local imports
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -43,8 +44,7 @@ import numpy as np
 def Coords2Int(coords: np.ndarray, spacing: np.ndarray, xmin: np.ndarray) -> np.ndarray:
     """ Compute the integer discretization in each direction
     """
-    disc = np.round((coords - xmin) * spacing)
-    return np.asarray(disc)
+    return np.round((coords - xmin) * spacing).astype(np.int64)
 
 
 def SFCResolution(kind: int, xmin: np.ndarray, xmax: np.ndarray) -> tuple[int, np.ndarray]:
@@ -53,7 +53,7 @@ def SFCResolution(kind: int, xmin: np.ndarray, xmax: np.ndarray) -> tuple[int, n
     """
     blen    = xmax - xmin
     nbits   = (kind*8 - 1)  # / 3.
-    intfact = cast(int, 2**nbits-1)
+    intfact = (1 << nbits) - 1
     spacing = np.ceil(intfact/blen)
 
     return np.ceil(nbits).astype(int), spacing
@@ -61,6 +61,8 @@ def SFCResolution(kind: int, xmin: np.ndarray, xmax: np.ndarray) -> tuple[int, n
 
 @final
 class tBox:
+    __slots__ = ('mini', 'intfact', 'spacing')
+
     def __init__(self, mini: int, maxi: int):
         self.mini = mini
         self.intfact = 0
@@ -71,10 +73,7 @@ class tBox:
         blen = maxi - mini
         nbits = (np.iinfo(np.int64).bits - 1) // 3
         self.intfact = 2 ** nbits - 1
-        if np.all(blen > 0):
-            self.spacing = self.intfact / blen
-        else:
-            self.spacing = self.intfact
+        self.spacing = np.where(blen > 0, self.intfact / blen, self.intfact)
 
 
 def SortMeshBySFC() -> None:
@@ -85,7 +84,6 @@ def SortMeshBySFC() -> None:
     from pyhope.common.common_progress import ProgressBar
     import pyhope.mesh.mesh_vars as mesh_vars
     import pyhope.output.output as hopout
-    import numpy as np
     # INFO: Alternative Hilbert curve sorting (not on PyPI)
     # from hilsort import hilbert_sort
     # ------------------------------------------------------
@@ -97,12 +95,13 @@ def SortMeshBySFC() -> None:
     sides = mesh_vars.sides
 
     totalElems = len(elems)
+    totalSides = len(sides)
     bar = ProgressBar(value=totalElems, title='│              Preparing Elements', length=33)
 
     # Global bounding box
     points = mesh.points
-    xmin = np.min(points, axis=0)
-    xmax = np.max(points, axis=0)
+    xmin = points.min(axis=0)
+    xmax = points.max(axis=0)
 
     # Calculate the element barycenters and associated element offsets
     elem_bary  = calc_elem_bary(elems)
@@ -117,7 +116,7 @@ def SortMeshBySFC() -> None:
     # Generate the space-filling curve and order elements along it
     hc             = HilbertCurve(p=nbits, n=3, n_procs=np_mtp)
 
[…]
