+++ tutorials/2-12-external_mesh_HOPR_merge/parameter.ini
@@ -14,6 +14,7 @@ FileName     = 2-12-external_mesh_HOPR_right_mesh.h5  ! Name of mesh file
 Mode         = 3                                    ! 1 Cartesian 3 External
 ! MeshScale    = 0.001                                ! Scale all input meshes by factor
 NGeo         = 4
+doSortIJK    = T                                    ! Sort the mesh elements along the I,J,K directions
 
 !================================================================================================================================= !
 ! BOUNDARY CONDITIONS
