### 2025-09-04 5611981
Perf: Improve performance of mesh output
