+++ pyhope/io/io.py
@@ -25,8 +25,8 @@
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
-from collections import OrderedDict
-from typing import Final
+from collections import defaultdict
+from typing import Final, Tuple, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 import h5py
@@ -89,20 +89,21 @@ def IO() -> None:
     hopout.separator()
     hopout.info('OUTPUT MESH...')
 
+    pname:  Final[str] = io_vars.projectname
+
     match io_vars.outputformat:
         case MeshFormat.FORMAT_HDF5:
             mesh  = mesh_vars.mesh
-            elems = mesh_vars.elems
-            sides = mesh_vars.sides
+            elems: Final[list] = cast(list, mesh_vars.elems)
+            sides: Final[list] = cast(list, mesh_vars.sides)
+            bcs:   Final[list] = cast(list, mesh_vars.bcs)
 
-            nElems = len(elems)
-            nSides = len(sides)
-            nNodes = sum(s.nodes.size for s in elems)  # number of non-unique nodes
+            nElems: Final[int] = len(elems)
+            nSides: Final[int] = len(sides)
+            # Number of non-unique nodes
+            nNodes: Final[int] = np.array([elem.nodes.size for elem in elems], dtype=np.int32).sum(dtype=int)
+            nBCs:   Final[int] = len(bcs)
 
-            bcs   = mesh_vars.bcs
-            nBCs  = len(bcs)
-
-            pname = io_vars.projectname
             fname = '{}_mesh.h5'.format(pname)
 
             elemInfo, sideInfo, nodeInfo, nodeCoords, \
@@ -112,9 +113,9 @@ def IO() -> None:
 
             # Print the final output
             hopout.sep()
-            for elemType in ELEM.TYPES:
-                if elemCounter[elemType] > 0:
-                    hopout.info( ELEMTYPE(elemType) + ': {:12d}'.format(elemCounter[elemType]))
+            elem_types = [(elemType, count) for elemType in ELEM.TYPES if (count := elemCounter.get(elemType, 0)) > 0]
+            for elemType, count in elem_types:
+                hopout.info(f'{ELEMTYPE(elemType)}: {count:12d}')
 
             hopout.sep()
             hopout.routine('Writing HDF5 mesh to "{}"'.format(fname))
@@ -153,10 +154,8 @@ def IO() -> None:
 
                 # Store boundary information
                 f.attrs['nBCs'          ] = nBCs
-                bcNames = [f'{s.name:<255}' for s in bcs]
-                bcTypes = np.zeros((nBCs, 4), dtype=np.int32)
-                for iBC, bc in enumerate(bcs):
-                    bcTypes[iBC, :] = bc.type
+                bcNames = [f'{bc.name:<255}' for bc in bcs]
+                bcTypes = np.array([bc.type  for bc in bcs], dtype=np.int32).reshape(-1, 4)  # noqa: E272
 
                 _ = f.create_dataset('BCNames'   , data=np.array(bcNames, dtype='S'))
                 _ = f.create_dataset('BCType'    , data=bcTypes)
@@ -164,14 +163,11 @@ def IO() -> None:
[…]
