### 2025-09-04 9a64d90
Perf: Improve readers, eliminate unnecessary copies
