+++ pyhope/mesh/reader/reader_hopr.py
@@ -25,15 +25,15 @@
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
-import copy
 import gc
 import itertools
 import os
 import shutil
+from collections import defaultdict
 # from dataclasses import dataclass, field
 from functools import cache
 from string import digits
-from typing import cast
+from typing import Any, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -89,7 +89,7 @@ def ReadHOPR(fnames: list, mesh: meshio.Mesh) -> meshio.Mesh:
     points   = mesh.points if len(mesh.points.shape)>1 else np.zeros((0, 3), dtype=np.float64)
     pointl   = cast(list, points.tolist())
     cells    = mesh.cells_dict
-    cellsets = {}
+    cellsets = defaultdict(lambda: defaultdict(list))
 
     nodeCoords   = mesh.points
     offsetnNodes = nodeCoords.shape[0]
@@ -178,8 +178,15 @@ def ReadHOPR(fnames: list, mesh: meshio.Mesh) -> meshio.Mesh:
             sideInfo   = np.array(f['SideInfo'])
             BCNames    = [s.strip().decode('utf-8') for s in cast(h5py.Dataset, f['BCNames'])]
 
-            # Cache the mapping here, so we consider the mesh order
-            linCache   = {}
+            # Pre-compute LINTEN mappings for all element types
+            # > Cache the mapping here, so we consider the mesh order
+            linCache  = {}
+            elemOrder = 100 if mesh_vars.nGeo == 1 else 200
+            elemTypes = tuple([s + elemOrder for s in (4, 5, 6, 8)])
+            for elemType in elemTypes:
+                _, mapLin = LINTEN(elemType, order=mesh_vars.nGeo)
+                mapLin    = np.array(tuple(mapLin[np.int64(i)] for i in range(len(mapLin))))
+                linCache[elemType] = mapLin
 
             with alive_bar(len(elemInfo), title='│             Processing Elements', length=33) as bar:
                 # Construct the elements, meshio format
@@ -192,24 +199,21 @@ def ReadHOPR(fnames: list, mesh: meshio.Mesh) -> meshio.Mesh:
                     elemType = elemTypeClass.inam[elemNum]
                     if len(elemType) > 1:
                         elemType  = elemType[0].rstrip(digits)
-                        elemType += str(NDOFperElemType(elemType, mesh_vars.nGeo))
+                        elemDOFs  = NDOFperElemType(elemType, mesh_vars.nGeo)
+                        elemType += str(elemDOFs)
                     else:
                         elemType  = elemType[0]
+                        elemDOFs  = NDOFperElemType(elemType, mesh_vars.nGeo)
 
                     # ChangeBasis currently only supported for hexahedrons
-                    if elemNum in linCache:
-                        mapLin = linCache[elemNum]
-                    else:
-                        _, mapLin = LINTEN(elemNum, order=mesh_vars.nGeo)
-                        mapLin    = np.array(tuple(mapLin[np.int64(i)] for i in range(len(mapLin))))
-                        linCache[elemNum] = mapLin
+                    mapLin = linCache[elemNum]
 
                     if nGeo == mesh_vars.nGeo:
                         elemIDs   = np.arange(elem[4], elem[5])
                         elemNodes = elemIDs[mapLin[:len(elemIDs)]]
[…]
