+++ pyhope/io/io.py
@@ -26,11 +26,12 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 from collections import defaultdict
-from typing import Final, cast
+from typing import Final, List, Iterable, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 import h5py
 import numpy as np
+from meshio import CellBlock
 # ----------------------------------------------------------------------------------------------------------------------------------
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Local imports
@@ -84,6 +85,7 @@ def IO() -> None:
     import pyhope.output.output as hopout
     from pyhope.common.common_vars import Common
     from pyhope.io.io_debug import DebugIO
+    from pyhope.io.io_gmsh import gmshEntityFromBlock, GMSHCELLTYPES
     from pyhope.io.io_vars import MeshFormat, ELEM, ELEMTYPE
     # ------------------------------------------------------
 
@@ -195,24 +197,101 @@ def IO() -> None:
             mesh  = mesh_vars.mesh
             fname = '{}_mesh.msh'.format(pname)
 
-            # Mixed elements required gmsh:dim_tags
-            # > FIXME: THIS ARE DUMMY ENTRIES AND ONLY GENERATE A POINT MESH
-            mesh.point_data.update({'gmsh:dim_tags': np.array([[0, i] for i in range(len(mesh.points))])})
+            # Instantiate the Gmsh cell type mapping
+            gmshCellTypes = GMSHCELLTYPES()
+
+            # Combine volume and surface cells as separate cell blocks
+            volume_cells  = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes3D]
+            surface_cells = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes2D]
+            # all_cells     = volume_cells + surface_cells
+
+            # Create new mesh with volume and surface elements
+            # volume_mesh = type(mesh)(points = mesh.points,  # noqa: E251
+            #                          cells  = all_cells)    # noqa: E251
+
+            # Build new point array with per-entity duplication (3D and 2D)
+            pointl:    List[np.ndarray] = []
+            pointt:    List[List[int] ] = []  # list of [dim, tag] rows parallel to pointl
+
+            celll:     List[CellBlock ] = []
+            celldphys: List[np.ndarray] = []
+            celldgeom: List[np.ndarray] = []
+
+            # Unique geometrical entity ids
+            geom_id:   np.ndarray       = np.zeros((2,), dtype=int)
+
+            # Process each 3D CellBlock: duplicate points, remap connectivity, set tags
+            for cell_block in volume_cells:
+                # TODO: Get the physical region from mesh_vars.elems
+                zone = 1
+
+                cb, phys, geom = gmshEntityFromBlock(points    = mesh.points,  # noqa: E251
+                                                     block     = cell_block,   # noqa: E251
+                                                     dim       = 3,            # noqa: E251
+                                                     geom_id   = geom_id[0],   # noqa: E251
+                                                     zones     = zone,         # noqa: E251
+                                                     uptPoints = pointl,       # noqa: E251
+                                                     uptDimTag = pointt)       # noqa: E251
+
+                # Update the lists
+                geom_id[0] += 1
+                celll    .append(cb)
+                celldphys.append(phys)
+                celldgeom.append(geom)
+
+            # Build a reverse mapping from the cell sets to the cell types
+            cell_type_to_bc   = {s.type: {} for s in surface_cells}
[…]
