### 2025-09-05 080ba8c
Output: Boundary conditions for Gmsh output
