+++ pyhope/io/io.py
@@ -26,7 +26,7 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 from collections import defaultdict
-from typing import Final, List, Iterable, cast
+from typing import Final, List, cast, Dict, Set
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 import h5py
@@ -85,7 +85,7 @@ def IO() -> None:
     import pyhope.output.output as hopout
     from pyhope.common.common_vars import Common
     from pyhope.io.io_debug import DebugIO
-    from pyhope.io.io_gmsh import gmshEntityFromBlock, GMSHCELLTYPES
+    from pyhope.io.io_gmsh import GMSHCELLTYPES
     from pyhope.io.io_vars import MeshFormat, ELEM, ELEMTYPE
     # ------------------------------------------------------
 
@@ -203,91 +203,155 @@ def IO() -> None:
             # Combine volume and surface cells as separate cell blocks
             volume_cells  = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes3D]
             surface_cells = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes2D]
-            # all_cells     = volume_cells + surface_cells
 
-            # Create new mesh with volume and surface elements
-            # volume_mesh = type(mesh)(points = mesh.points,  # noqa: E251
-            #                          cells  = all_cells)    # noqa: E251
+            # Build new arrays
+            celll:      List[CellBlock ] = []
+            celldphys:  List[np.ndarray] = []
+            celldgeom:  List[np.ndarray] = []
 
-            # Build new point array with per-entity duplication (3D and 2D)
-            pointl:    List[np.ndarray] = []
-            pointt:    List[List[int] ] = []  # list of [dim, tag] rows parallel to pointl
+            # Unique geometrical entity ids per dimension
+            # > 0: 3D entities, 1: 2D entities
+            geom_id:    np.ndarray       = np.ones((2,), dtype=int)
+            geom_tag:   List[List[int]]  = [[] for _ in range(2)]
+            geom_nodes: List[List[int]]  = [[] for _ in range(2)]
 
-            celll:     List[CellBlock ] = []
-            celldphys: List[np.ndarray] = []
-            celldgeom: List[np.ndarray] = []
+            # Set to keep track of nodes already used to represent an entity in gmsh:dim_tags
+            usedNodes: Set[int] = set()
 
-            # Unique geometrical entity ids
-            geom_id:   np.ndarray       = np.zeros((2,), dtype=int)
-
-            # Process each 3D CellBlock: duplicate points, remap connectivity, set tags
+            # Process each 3D CellBlock: keep shared connectivity, set tags
+            # WARNING: Each 3D CellBlock neets to get its OWN geometrical tag so that
+            #          meshio._write_entities() does not see duplicate (dim, tag) across cell blocks
             for cell_block in volume_cells:
                 # TODO: Get the physical region from mesh_vars.elems
                 zone = 1
 
-                cb, phys, geom = gmshEntityFromBlock(points    = mesh.points,  # noqa: E251
-                                                     block     = cell_block,   # noqa: E251
-                                                     dim       = 3,            # noqa: E251
-                                                     geom_id   = geom_id[0],   # noqa: E251
-                                                     zones     = zone,         # noqa: E251
-                                                     uptPoints = pointl,       # noqa: E251
-                                                     uptDimTag = pointt)       # noqa: E251
+                # Shared nodes, so keep the original connectivity
+                cb     = CellBlock(cell_block.type, cell_block.data.copy())
[…]
