### 2025-09-05 13505c6
Output: Add the final element count to Gmsh
