### 2025-09-05 9a10d0d
Output: Refactor for generalized meshio_to_gmsh routine
