+++ pyhope/io/io.py
@@ -26,12 +26,11 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 from collections import defaultdict
-from typing import Dict, Final, List, Set, cast
+from typing import Final, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 import h5py
 import numpy as np
-from meshio import CellBlock
 # ----------------------------------------------------------------------------------------------------------------------------------
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Local imports
@@ -87,6 +86,7 @@ def IO() -> None:
     from pyhope.io.io_debug import DebugIO
     from pyhope.io.io_gmsh import GMSHCELLTYPES
     from pyhope.io.io_vars import MeshFormat, ELEM, ELEMTYPE
+    from pyhope.meshio.meshio_convert import meshio_to_gmsh
     from pyhope.meshio.meshio_nodes import NumNodesPerCell
     # ------------------------------------------------------
 
@@ -205,178 +205,28 @@ def IO() -> None:
             # Instantiate the Gmsh cell type mapping
             gmshCellTypes = GMSHCELLTYPES()
 
-            # Combine volume and surface cells as separate cell blocks
-            volume_cells  = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes3D]
-            surface_cells = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes2D]
-
             # Print the final output
             hopout.sep()
             numNodes = NumNodesPerCell()
-            for cell in volume_cells:
+            for cell in [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes3D]:
                 cellType  = ''.join([s for s in cell.type if not s.isdigit()])
                 cellNodes = numNodes[cellType]
                 elemOrder = 100 if not any(s.isdigit() for s in cell.type) else 200
                 elemType  = cellNodes + elemOrder
                 hopout.info(f'{ELEMTYPE(elemType)}: {len(cell):12d}')
 
-            # Build new arrays
-            celll:      List[CellBlock ] = []
-            celldphys:  List[np.ndarray] = []
-            celldgeom:  List[np.ndarray] = []
+            gmshMesh = meshio_to_gmsh(mesh)
 
-            # Unique geometrical entity ids per dimension
-            # > 0: 3D entities, 1: 2D entities
-            geom_id:    np.ndarray       = np.ones((2,), dtype=int)
-            geom_tag:   List[List[int]]  = [[] for _ in range(2)]
-            geom_nodes: List[List[int]]  = [[] for _ in range(2)]
-
-            # Set to keep track of nodes already used to represent an entity in gmsh:dim_tags
-            usedNodes: Set[int] = set()
-
-            # Process each 3D CellBlock: keep shared connectivity, set tags
-            # WARNING: Each 3D CellBlock neets to get its OWN geometrical tag so that
-            #          meshio._write_entities() does not see duplicate (dim, tag) across cell blocks
-            for cell_block in volume_cells:
-                # TODO: Get the physical region from mesh_vars.elems
-                zone = 1
-
-                # Shared nodes, so keep the original connectivity
-                cb     = CellBlock(cell_block.type, cell_block.data.copy())
-                nCells = len(cell_block.data)
-                # Unique 3D geometrical tag for volume block
-                tag    = int(geom_id[0])
-                phys   = np.full(nCells, int(zone), dtype=int)
-                geom   = np.full(nCells, tag,       dtype=int)
-
[…]
