+++ pyhope/meshio/meshio_convert.py
@@ -26,6 +26,7 @@
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
+from typing import Dict, Final, List, Set, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -102,3 +103,169 @@ def gmsh_to_meshio(gmsh) -> meshio.Mesh:
         cell_sets[name] = [(None if len(idcs) == 0 else np.concatenate(idcs)) for idcs in cell_sets[name]]
 
     return meshio.Mesh(points, cells, cell_sets=cell_sets)
+
+
+def meshio_to_gmsh(mesh: meshio.Mesh) -> meshio.Mesh:
+    # Local imports ----------------------------------------
+    import pyhope.mesh.mesh_vars as mesh_vars
+    import pyhope.output.output as hopout
+    from pyhope.io.io_gmsh import GMSHCELLTYPES
+    # ------------------------------------------------------
+
+    # Instantiate the Gmsh cell type mapping
+    gmshCellTypes = GMSHCELLTYPES()
+
+    # Combine volume and surface cells as separate cell blocks
+    volume_cells  = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes3D]
+    surface_cells = [cell_block for cell_block in mesh.cells if cell_block.type in gmshCellTypes.cellTypes2D]
+
+    # Build new arrays
+    celll:      List[meshio.CellBlock] = []
+    celldphys:  List[np.ndarray      ] = []
+    celldgeom:  List[np.ndarray      ] = []
+
+    # Unique geometrical entity ids per dimension
+    # > 0: 3D entities, 1: 2D entities
+    geom_id:    np.ndarray       = np.ones((2,), dtype=int)
+    geom_tag:   List[List[int]]  = [[] for _ in range(2)]
+    geom_nodes: List[List[int]]  = [[] for _ in range(2)]
+
+    # Set to keep track of nodes already used to represent an entity in gmsh:dim_tags
+    usedNodes: Set[int] = set()
+
+    # Process each 3D CellBlock: keep shared connectivity, set tags
+    # WARNING: Each 3D CellBlock neets to get its OWN geometrical tag so that
+    #          meshio._write_entities() does not see duplicate (dim, tag) across cell blocks
+    for cell_block in volume_cells:
+        # TODO: Get the physical region from mesh_vars.elems
+        zone = 1
+
+        # Shared nodes, so keep the original connectivity
+        cb     = meshio.CellBlock(cell_block.type, cell_block.data.copy())
+        nCells = len(cell_block.data)
+        # Unique 3D geometrical tag for volume block
+        tag    = int(geom_id[0])
+        phys   = np.full(nCells, int(zone), dtype=int)
+        geom   = np.full(nCells, tag,       dtype=int)
+
+        # Choose a representative node for this 3D entity (unique across all entities)
+        node_cand = cast(np.ndarray, cb.data).ravel().astype(int)
+        node_free = [s for s in node_cand if s not in usedNodes]
+        node_used = int(node_free[0] if node_free else node_cand[0])
+        if not node_free and node_used in usedNodes:
+            # All candidate nodes already used; fall back to the first (still valid, we just can't make it unique)
+            # hopout.routine(f'Note: reusing representative node {rep_node} for 3D entity tag {tag}')
[…]
