+++ tutorials/1-02-cartbox_periodic/parameter.ini
@@ -3,7 +3,7 @@
 !================================================================================================================================= !
 ProjectName  = 1-02-cartbox_periodic                ! Name of output files
 DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
-OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
+OutputFormat = GMSH                                 ! Mesh output format (HDF5 VTK)
 DebugMesh    = T                                    ! Output debug mesh in XDMF format
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
