### 2025-09-05 6c13c6f
Bugfix: Explicitly request OpenMP multithreading in Gmsh
