+++ pyhope/mesh/reader/reader_gmsh.py
@@ -89,6 +89,7 @@ def ReadGMSH(fnames: list) -> meshio.Mesh:
     import pyhope.mesh.mesh_vars as mesh_vars
     import pyhope.output.output as hopout
     from pyhope.common.common import IsDisplay
+    from pyhope.common.common_vars import np_mtp
     from pyhope.io.io_vars import debugvisu
     from pyhope.mesh.topology.mesh_serendipity import convertSerendipityToFullLagrange
     from pyhope.meshio.meshio_convert import gmsh_to_meshio
@@ -97,7 +98,16 @@ def ReadGMSH(fnames: list) -> meshio.Mesh:
 
     hopout.sep()
     gmsh.initialize()
+
+    # Setup multiprocessing
+    numThreads = np_mtp if np_mtp > 0 else 1
+    gmsh.option.setNumber('General.NumThreads',   numThreads)
+    gmsh.option.setNumber('Geometry.OCCParallel', 1 if np_mtp > 0 else 0)
+
+    # Setup mesh factory
     # gmsh.option.setString('SetFactory', 'OpenCascade')
+
+    # Setup debug visualization
     if not debugvisu:
         # Hide the GMSH debug output
         gmsh.option.setNumber('General.Terminal', 0)
