### 2025-09-09 0b6b8fb
meshio: Remove intermediate meshio.Mesh
