### 2025-09-15 d475ba6
Gmsh: Update to v4.14.1
Start preparing to replace deprecated setuptools with uv
