### 2025-09-17 #8769 @jwallwork23
> [@jwallwork23](https://github.com/jwallwork23) How is your work on the open issues going?

The issues that I opened have associated pull requests but they haven't been reviewed yet. I've not had a look at the paper because I'm not sure if it's been updated?
