### 2025-09-19 hopr-framework/PyHOPE#3 @kopperp
Fixed in #15
