### 2025-09-19 55b1604
Bugfix: GitHub Actions to directly deploy pages
https://github.com/squidfunk/mkdocs-material/discussions/7038
