### 2025-09-19 bbb0e56
Documentation: Add contributing, code of conduct
