+++ CODE_OF_CONDUCT.md
@@ -1,5 +1,6 @@
-
 # Code of Conduct
+
+Contributor Covenant 3.0 Code of Conduct
 
 ## Our Pledge
 
