+++ pyproject.toml
@@ -1,7 +1,7 @@
 # pyporject.toml compatible with Poetry (version 2.0.0.dev0)
 [project]
 name = "PyHOPE"
-version = "0.1.0"
+version = "0.1.1"
 description = "Python High-Order Preprocessing Environment"
 license = {text = "GPL-3.0-only"}
 requires-python = ">=3.10"
