### 2025-09-22 741fb1e
Bugfix: Remove unused writeVTM function
