### 2025-09-23 322aa21
Improvement: Error message for invalid parameters
