+++ pyhope/readintools/readintools.py
@@ -63,16 +63,16 @@ def strToBool(val: Union[int, bool, str]) -> bool:  # From distutils.util.strtob
         False values are 'n', 'no' , 'f', 'false', 'off', and '0'.
         Raises ValueError if 'val' is anything else.
     """
-    if type(val) is bool:
+    if isinstance(val, bool):
         return val
-    if type(val) is int:
+    if isinstance(val, int):
         val = str(val)
-    if type(val) is str:
+    if isinstance(val, str):
         val = val.lower()
 
-    if val in ('y', 'yes', 't', 'true', 'on', '1'):
+    if   val in ('y', 'yes', 't', 'true' , 'on' , '1'):  # noqa: E271
         return True
-    elif val in ('n', 'no', 'f', 'false', 'off', '0'):
+    elif val in ('n', 'no' , 'f', 'false', 'off', '0'):  # noqa: E271
         return False
     else:
         raise ValueError('invalid truth value %r' % (val,))
@@ -316,7 +316,7 @@ def CountOption(string: str) -> int:
 def GetParam(name    : str,
              calltype: str,
              default : Optional[str] = None,
-             number  : Optional[int] = None):
+             number  : Optional[int] = None) -> str:
     # Local imports ----------------------------------------
     import pyhope.config.config as config
     import pyhope.output.output as hopout
@@ -397,20 +397,23 @@ def GetIntFromStr(name: str, default: Optional[str] = None, number: Optional[int
 
     # Check if we already received the int. Otherwise, get the value from the mapping
     mapping = config.prms[name]['mapping']
-    if type(value) is int:
-        hopout.printoption(name, '{} [{}]'.format(value, mapping[value]), source)
-    else:
-        if not value.isdigit():
-            value = [s for s, v in mapping.items() if v.lower() == value.lower()]
-            if len(value) == 0:
-                hopout.error('Unknown value for parameter {}, exiting...'.format(name), traceback=True)
-            else:
-                value = int(value[0])
-                hopout.printoption(name, '{} [{}]'.format(value, mapping[value]), source)
-        else:
-            hopout.printoption(name, '{} [{}]'.format(value, mapping[int(value)]), source)
+    options = {v.lower(): int(k) for k, v in mapping.items()}
 
-    return int(value)
+    result = None
+    try:
+        result = int(value)
+    except (ValueError, TypeError):
+        result = options.get(str(value).lower())
+
+    if result is None or result not in mapping.keys():  # pragma: no cover
+        outStr = ', '.join([f'{k} [{v}]' for k, v in mapping.items()])
+        print()
+        print(hopout.warn(f'Allowed values for parameter "{name}":'))
+        print(hopout.warn(f'{outStr}'))
+        hopout.error(f'Unknown value "{value}" for parameter "{name}", exiting...')
+
+    hopout.printoption(name, '{} [{}]'.format(result, mapping[result]), source)
+    return result
 
 
 def GetRealArray(name: str, default: Optional[str] = None, number: Optional[int] = None) -> np.ndarray:
@@ -430,10 +433,11 @@ def GetRealArray(name: str, default: Optional[str] = None, number: Optional[int]
         value = value.split(',')
     try:
         value = np.vectorize(strToFloatOrPi)(value)
-    except ValueError as e:
+    except ValueError as e:  # pragma: no cover
         print()
         print(hopout.warn(f'{e}'))
         hopout.error(f'Failed to read "{name}" array, possibly malformed comma-separated data. Exiting...')
+
     CheckDimension(name, value.size)
     return value
 
@@ -486,25 +490,36 @@ class ReadConfig():
                 # Remove all whitespaces
                 line = ''.join(line.split())
 
+                # Skip empty lines early
+                if not line:
+                    continue
+
[…]
