### 2025-09-25 5593a97
JOSS Manuscript: Incorporate review feedback by Joe Wallwork
