+++ paper.md
@@ -57,71 +57,73 @@ A crucial aspect of DGSEM is the mapping of the equations to be solved from refe
 At the same time, DGSEM features a highly local stencil since grid elements are connected solely via the numerical flux through adjacent element faces.
 Providing mesh and solution data in an on-disc data format which facilitates non-overlapping I/O is key to efficient parallel data access, thereby minimizing execution time.
 While HOPR has traditionally served as a reference implementation of a mesh generator for this process, PyHOPE is a modern alternative that enhances readability and extensibility.
-Designed with excellent scaling and parallel computing capabilities in mind, PyHOPE offers a more user-friendly and adaptable solution for researchers and engineers working on high-order mesh generation and transformation.
+Designed with a clear code structure and modularization in mind, PyHOPE offers a more user-friendly and adaptable solution for researchers and engineers working on high-order mesh generation and transformation.
 
 [^1]: [https://numericsresearchgroup.org/codes.html](https://numericsresearchgroup.org/codes.html)
 [^2]: [https://github.com/piclas-framework/piclas](https://github.com/piclas-framework/piclas)
 
+# State of the Field
+PyHOPE shares similarities with other high-order mesh generation tools such as HOPR which serves as the reference implementation.
+Although the HOPR format is supported by a variety of codes, there is limited code support to generate meshes in this format.
+Compared to HOPR, PyHOPE places a special focus on enhanced adaptability, including powerful element splitting functionality, improved usability, and broader support for modern mesh formats, including curved and mixed meshes to enable more complex mesh generation.
+The following spectral element solvers have (optional) support for meshes generated in PyHOPE.
+
++--------------+-------------+-------------------------+------------------------------------------+
+| Framework    | Language    | Equation System         | Reference                                |
++:=============+:============+:========================+:=========================================+
+| FLEXI        | Fortran     | NSE                     | [@Krais2021]                             |
++--------------+-------------+-------------------------+------------------------------------------+
+| ƎLEXI        | Fortran     | NSE/MRG                 | [@Kopper2023]                            |
++--------------+-------------+-------------------------+------------------------------------------+
+| GALÆXI       | Fortran/C   | NSE                     | [@Kurz2025]                              |
++--------------+-------------+-------------------------+------------------------------------------+
+| FLUXO        | Fortran     | NSE/MHD/Maxwell         | [@RuedaRamirez2017]                      |
++--------------+-------------+-------------------------+------------------------------------------+
+| HORSES3D     | Fortran     | NSE/Cahn-Hilliard       | [@Ferrer2023]                            |
++--------------+-------------+-------------------------+------------------------------------------+
+| PICLas       | Fortran     | Maxwell/Poisson         | [@Fasoulas2019]                          |
++--------------+-------------+-------------------------+------------------------------------------+
+| SELF         | Fortran     | Shallow Water/Euler     | github.com/FluidNumerics/SELF            |
++:=============+:============+:========================+:=========================================+
+|~*Equation Systems: NSE - Navier-Stokes, MRG - Maxey-Riley-Gatignol, MHD - Magnetohydrodynamics*~|
++:=============+:============+:========================+:=========================================+
+
 # Features
-PyHOPE facilitates high-order mesh generation and transformation through a streamlined and accessible interface.
+PyHOPE provides a simple and intuitive interface for generating and transforming high-order meshes.
 PyHOPE is distributed as open-source software on GitHub[^3] and as package on the Python Package Index (PyPI)[^4].
-It reads user input from a single plain-text configuration file in INI format, rendering setup and execution straightforward.
[…]
