### 2025-09-25 #8769 @kopperp
I have incorporated the code issues, please see
- [[JOSS review]: Community guidelines](https://github.com/hopr-framework/PyHOPE/issues/3) - [PR #15](https://github.com/hopr-framework/PyHOPE/pull/15)
- [[JOSS review]: Functionality documentation](https://github.com/hopr-framework/PyHOPE/issues/4) - [PR #15](https://github.com/hopr-framework/PyHOPE/pull/15)
- [[JOSS review]: Automated tests](https://github.com/hopr-framework/PyHOPE/issues/5) - [PR 6](https://github.com/hopr-framework/PyHOPE/pull/6), [PR 11](https://github.com/hopr-framework/PyHOPE/pull/11)

The changes following the feedback of the paper are in [PR 37](https://github.com/hopr-framework/PyHOPE/pull/37). I have spun off the points into individual comments in an attempt to structure the discussion. If @jwallwork23 agrees with the changes, I will merge them into the JOSS paper branch.
