### 2025-09-25 9923256
Perf: Improve mesh checks
