+++ tutorials/3-03-sec_order_mixed_CGNS/parameter.ini
@@ -4,7 +4,7 @@
 ProjectName  = 3-03-sec_order_mixed_CGNS            ! Name of output files
 DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
-nThreads     = 4                                    ! Number of threads for multiprocessing
+! nThreads     = 4                                    ! Number of threads for multiprocessing
 
 !=============================================================================== !
 ! MESH
