### 2025-09-26 3b4e9e6
DebugMesh: Fix missing integer conversion for Gambit meshes
