### 2025-09-26 f2bdf5f
CI/CD: Enable DebugMesh for external mesh formats
