+++ tutorials/2-03-external_mesh_Gmsh_curved/parameter.ini
@@ -3,6 +3,7 @@
 !=============================================================================== !
 ProjectName  = 2-03-external_mesh_Gmsh_curved       ! Name of output files
 DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugMesh    = T                                    ! Mesh output debug mesh in VTK format
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
