+++ tutorials/1-11-cartbox_mixed/parameter.ini
@@ -2,8 +2,8 @@
 ! OUTPUT
 !================================================================================================================================= !
 ProjectName  = 1-11-cartbox_mixed                   ! Name of output files
-DebugMesh    = T
 DebugVisu    = F                                    ! Launch the GMSH GUI to visualize the mesh
+DebugMesh    = T                                    ! Mesh output debug mesh in VTK format
 OutputFormat = HDF5                                 ! Mesh output format (HDF5 VTK)
 ! nThreads     = 4                                    ! Number of threads for multiprocessing
 
