### 2025-09-26 77b0f58
DebugMesh: Add FEM vertex information output
