### 2025-10-01 d93be92
introduced snake-like and lexicographic element ordering (not called anywhere though)
