### 2025-10-03 #8769 @mohd-afeef-badri
@kopperp and co-authors,

I have completed my review of the PyHOPE package. Please find my comments below:

- [ ]  **Installation and tutorials**: I had no issues following your installation instructions. The inclusion of a `health-check` step post-installation, as well as the automated check/update of Gmsh with the correct CGNS version, is very much appreciated. This greatly improves the first-time user experience, at least it did for me. Post-installation I was able to check the tutorials run without issue

```
✅ OK Successfully completed 22/22 tests
```
 However, when I attempted to run these tests by myself, most of them ran, but the CGNS ones, e.g, `3-05-sec_order_layers_CGNS`, does not run with errors

```
Exception: CGNS error cgio_open_file:invalid file type
```

this has indeed got to do with perhaps wrong Gmsh version being used, but what I fail to understand why does this test run when it is ran as part of `pyhope --verify` as I see

```
│ ✅ OK Successfully completed test "3-05-sec_order_layers_CGNS"
```

but it does not run when I run it manually
```
cd tutorials/3-05-sec_order_layers_CGNS && pyhope ./parameter.ini
```

- [ ] **FEM support**: Via the `--help` menu I  could see that *pyHOPE* appears to include FEM support (if constructing FEM connectivity is requested). Could you clarify if this is indeed the case? If so, it would be helpful to mention this explicitly in the paper.

- [ ] **Documentation rendering issue**: In your documentation (`PyHOPE/docs/getting-started.md`), the final code block does not render correctly on the GitHub Pages site. Please fix this for clarity.

- [ ] **Workflows and tutorials**: The two workflows on how to use PyHOPE via [internal mesh generator](https://hopr-framework.github.io/PyHOPE/user-guide/mesh-generators/internal/) or via [external mesh generators](https://hopr-framework.github.io/PyHOPE/user-guide/mesh-generators/external/) are clearly explained. However, it would be very useful to include example images of meshes corresponding to each workflow. While some readers may prefer to imagine the mesh structure from the text, others will benefit from visual references.

- [ ] **Internal mesh output format**: In the [internal mesher documentation](https://hopr-framework.github.io/PyHOPE/user-guide/mesh-generators/internal/), setting `OutputFormat` to HDF5 seems to produce the HOPR mesh format, while the text also mentions that VTK can be used. Could you clarify this point? Does VTK include higher order mesh support ?

- [ ] **External mesh formats**: In the [external mesher tutorial](https://hopr-framework.github.io/PyHOPE/user-guide/mesh-generators/external/), you note that files can be read through Gmsh or internal readers. However, the table only lists `.msh` and CGNS formats, even though Gmsh is capable of handling other formats (VTK, MED, etc.). Could this be clarified? can other mesh formats be used as inputs ?

- [ ] **Python library usage**: The [Getting Started](https://hopr-framework.github.io/PyHOPE/getting-started/) page makes it clear that *pyHOPE* can be used via both CLI and as a Python library. However, the User Guide (tutorial section) does not provide any example of using the Python API. It would be valuable to include a simple example showing how to construct a mesh using the library interface, since the API documentation is missing.

Thank you in advance for your replies and efforts.
