### 2025-10-06 949f9e5
DebugMesh: Move to XDMF, add FEM edge information output
