+++ pyhope/io/io_xdmf.py
@@ -25,13 +25,16 @@
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
-import os
-from functools import cache
+import pathlib
+from typing import Union
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 import h5py
-import numpy as np
+from meshio import Mesh
+from meshio._common import write_xml
+from meshio._exceptions import WriteError
+from xml.etree import ElementTree as ET
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Local imports
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -41,150 +44,203 @@ import numpy as np
 # ==================================================================================================================================
 
 
-@cache
-def xdmfElems(elemType: int, nGeo: int) -> tuple[str, int]:
-    """ XDMF: Get the element type from the HOPR format
-    - Linear
-        - Polyvertex - a group of unconnected points
-        - Polyline - a group of line segments
-        - Polygon
-        - Triangle
-        - Quadrilateral
-        - Tetrahedron
-        - Pyramid
-        - Wedge
-        - Hexahedron
-    - Quadratic
-        - Edge_3 - Quadratic line with 3 nodes
-        - Tri_6
-        - Quad_8
-        - Tet_10
-        - Pyramid_13
-        - Wedge_15
-        - Hex_20
-    - Arbitrary
-        - 1 - POLYVERTEX
-        - 2 - POLYLINE
-        - 3 - POLYGON
-        - 4 - TRIANGLE
-        - 5 - QUADRILATERAL
-        - 6 - TETRAHEDRON
-        - 7 - PYRAMID
-        - 8 - WEDGE
-        - 9 - HEXAHEDRON
-        - 16 - POLYHEDRON
-        - 34 - EDGE_3
-        - 35 - QUADRILATERAL_9
-        - 36 - TRIANGLE_6
-        - 37 - QUADRILATERAL_8
-        - 38 - TETRAHEDRON_10
-        - 39 - PYRAMID_13
-        - 40 - WEDGE_15
-        - 41 - WEDGE_18
-        - 48 - HEXAHEDRON_20
-        - 49 - HEXAHEDRON_24
-        - 50 - HEXAHEDRON_27
-    """
-    elem_map = { 4: 'TETRAHEDRON',
-                 5: 'PYRAMID',
-                 6: 'WEDGE',
-                 8: 'HEXAHEDRON'
-               }
+# Monkey-patching meshio.xdmf.main.XdmfWriter
+# > We want to be able to write multiple grids into the same HDF5 file
+def XdmfWriterInit(self,
+                   filename        : str,
+                   meshes          : Union[Mesh, list[Mesh]],
+                   data_format     : str = 'HDF',
+                   compression     : str = 'gzip',
+                   compression_opts: int = 4) -> None:
 
-    if elemType % 10 not in elem_map:
-        raise ValueError(f'Error in face_to_edge: elemType {elemType} is not supported')
+    if data_format not in ['XML', 'Binary', 'HDF']:
[…]
