### 2025-10-06 0592473
CI/CD: Reduce Python package installations
