### 2025-10-06 d1d030d
Check: Add consistency checks (BCs, Stretching)
