### 2025-10-06 5b4c687
Check: Add consistency check (file size)
