### 2025-10-07 cce99c1
Feature: Mesh sorting (SFC, IJK, LEX, Snake)
