+++ pyhope/mesh/mesh_sort.py
@@ -26,7 +26,7 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 import gc
-from typing import Final, cast, final
+from typing import Final, List, Optional, Tuple, cast, final
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
@@ -57,6 +57,57 @@ def SFCResolution(kind: int, xmin: np.ndarray, xmax: np.ndarray) -> tuple[int, n
     spacing = np.ceil(intfact/blen)
 
     return np.ceil(nbits).astype(int), spacing
+
+
+def UpdateElemID(elems         : list,
+                 sides         : list,
+                 sorted_indices: np.ndarray,
+                 bar,
+                 nElemsIJK     : Optional[np.ndarray] = None,
+                 ) -> Tuple[List, List]:
+
+    totalElems = len(elems)
+    totalSides = len(sides)
+
+    # Initialize sorted cells
+    sorted_elems = [None] * totalElems
+    sorted_sides = [None] * totalSides
+
+    bar.title('│             Processing Elements')
+
+    # Initialize the sideID and offset
+    offsetSide = 0
+    sideID     = 0
+
+    # Overwrite the elem/side IDs
+    for newElemID, oldElemID in enumerate(sorted_indices):
+        elem        = elems[oldElemID]
+        elem.elemID = newElemID
+
+        # Calculate IJK position for this element
+        if nElemsIJK is not None:
+            k =  newElemID                                       // (nElemsIJK[0] * nElemsIJK[1])
+            j = (newElemID - k * nElemsIJK[0] * nElemsIJK[1])    //  nElemsIJK[0]
+            i =  newElemID - k * nElemsIJK[0] * nElemsIJK[1] - j *   nElemsIJK[0]
+            elem.elemIJK = np.array([i+1, j+1, k+1], dtype=np.int32)
+
+        sorted_elems[newElemID] = elem
+
+        # Correct the sideID
+        for key, val in enumerate(elem.sides):
+            side        = sides[val]
+            side.sideID = offsetSide + key
+            side.elemID = newElemID
+            sorted_sides[sideID] = side
+            sideID     += 1
+
+        # Correct the sideID
+        nSides      = len(elem.sides)
+        elem.sides  = list(range(offsetSide, offsetSide + nSides))
+        offsetSide += nSides
+        bar.step()
+
+    return sorted_elems, sorted_sides
 
 
 @final
@@ -94,9 +145,7 @@ def SortMeshBySFC() -> None:
     elems = mesh_vars.elems
     sides = mesh_vars.sides
 
-    totalElems = len(elems)
-    totalSides = len(sides)
-    bar = ProgressBar(value=totalElems, title='│              Preparing Elements', length=33)
+    bar = ProgressBar(value=len(elems), title='│              Preparing Elements', length=33)
 
     # Global bounding box
     points = mesh.points
@@ -126,35 +175,7 @@ def SortMeshBySFC() -> None:
     # Now, create an array that maps each element to the new sorting
     # sorted_indices = np.array([value_to_index[tuple(val.tolist())] for val in elem_bary])
 
-    # Initialize sorted cells
-    sorted_elems = [None] * totalElems
-    sorted_sides = [None] * totalSides
-
-    bar.title('│             Processing Elements')
-
-    # Initialize the sideID and offset
-    offsetSide = 0
-    sideID     = 0
-
-    # Overwrite the elem/side IDs
[…]
