+++ tutorials/2-12-external_mesh_HOPR_merge/parameter_left.ini
@@ -11,7 +11,8 @@ OutputFormat = HDF5                                 ! Mesh output format (HDF5 V
 !================================================================================================================================= !
 Mode         = 1                                    ! Mode for Cartesian boxes
 nZones       = 1                                    ! Number of boxes
-doSortIJK    = T                                    ! Sort the mesh elements along the I,J,K directions
+MeshSorting  = IJK                                  ! Sort the mesh elements along the I,J,K directions
+! doSortIJK    = T                                    ! Sort the mesh elements along the I,J,K directions
 
 ! ===  left zone  ===
 Corner       = (/-1,-1,-1   ,,0,-1,-1   ,,0,1,-1    ,,-1,1,-1    ,,-1,-1,1   ,,0,-1,1   ,,0,1,1    ,,-1,1,1/)
