### 2025-10-07 fc6ac76
Linting: Add type checking with ty, fix issues
