### 2025-10-08 b9321be
Documentation: Clarify FEM connectivity
