### 2025-10-08 3692a28
Documentation: Add Git LFS to README.md
