### 2025-10-08 5bd5436
Documentation: Add mesh example images
