### 2025-10-08 76e1530
Documentation: Add usage as Python library
