+++ mkdocs.yml
@@ -44,6 +44,7 @@ nav:
   - User Guide:
     - Introduction: user-guide/index.md
     - Installation: user-guide/installation.md
+    - Library Usage: user-guide/library_usage.md
     - Mesh Generators:
       - Internal: user-guide/mesh-generators/internal.md
       - External: user-guide/mesh-generators/external.md
