### 2025-10-10 7bf4d94
Documentation: Add function signatures to library documentation
