+++ pyhope/basis/basis_basis.py
@@ -148,7 +148,7 @@ def change_basis_3D(Vdm: np.ndarray, x3D_In: np.ndarray) -> np.ndarray:
     """ Interpolate a 3D tensor product Lagrange basis defined by (N_in+1) 1D interpolation point positions xi_In(0:N_In)
         to another 3D tensor product node positions (number of nodes N_out+1)
         defined by (N_out+1) interpolation point  positions xi_Out(0:N_Out)
-        xi is defined in the 1DrefElem xi=[-1,1]
+        xi is defined in the 1D reference element xi=[-1,1]
     """
     # First contraction along the iN_In axis (axis 1 of Vdm, axis 1 of x3D_In)
     x3D_Buf1 = np.tensordot(Vdm, x3D_In , axes=(1, 1))
@@ -187,6 +187,8 @@ def change_basis_2D(Vdm: np.ndarray, x2D_In: np.ndarray) -> np.ndarray:
 
 
 def evaluate_jacobian(xGeo_In: np.ndarray, VdmGLtoAP: np.ndarray, D_EqToGL: np.ndarray) -> np.ndarray:
+    """ Calculate the Jacobian of the mapping for a given element
+    """
     # Perform tensor contraction for the first derivative (Xi direction)
     dXdXiGL   = np.tensordot(D_EqToGL, xGeo_In, axes=(1, 1))
     dXdXiGL   = np.moveaxis(dXdXiGL  , 1, 0)  # Correct the shape to (3, nGeoRef, nGeoRef, nGeoRef)
