### 2025-10-10 b1b7056
Bugfix: Add missing IJK arrays
