### 2025-10-10 f8ad266
Linting: Explicit type hints
