### 2025-10-10 bcbb19e
JOSS Manuscript: Incorporate review feedback by Mohd Afeef Badri
