+++ paper.md
@@ -92,7 +92,7 @@ The following spectral element solvers have (optional) support for meshes genera
 PyHOPE provides a simple and intuitive interface for generating and transforming high-order meshes.
 PyHOPE is distributed as open-source software on GitHub[^3] and as package on the Python Package Index (PyPI)[^4].
 It reads user input from a single plain-text configuration file in INI format[^5].
-The software supports the generation of block‑structured meshes using canonical volumetric element types (tetrahedral, pyramidal, prismatic, and hexahedral), allowing for mesh stretching and post‑deformation at arbitrary polynomial orders.
+The software supports the generation of block‑structured meshes using any combination of canonical volumetric element types (tetrahedral, pyramidal, prismatic, and hexahedral), allowing for mesh stretching and post‑deformation at arbitrary polynomial orders.
 PyHOPE can automatically create rectilinear boundary layer meshes based on a desired wall resolution or stretching factor.
 It can read in and merge both internally and externally created curved meshes while incorporating boundary conditions and automatically adapting sub-meshes to the desired polynomial order.
 PyHOPE also offers conversion of simplex elements into fully hexagonal cells through geometric subdivision, as well as mesh sorting along structured dimensions or space-filling Hilbert curves.
@@ -113,13 +113,12 @@ The available post-deformation options and topology conversion features are outl
 All tutorial cases are also used for code coverage and regression checking during Continuous Integration/Continuous Deployment (CI/CD).
 
 One notable example from the field of aerospace engineering is the application of PyHOPE to the Common Research Model (CRM), a widely used aerodynamic benchmark.
-For this example, the CAD file is initially meshed using second-order simplex elements in the external generator ANSA v24 and exported in high-order curved CGNS 4.2.0 format based on HDF5.
+For this example, the CAD file is initially meshed using second-order simplex elements in the external generator ANSA v24 [@ansa2024] and the three-dimensional grid is exported in high-order curved CGNS 4.2.0 format based on HDF5.
 While the specific mesh used here does not include boundary layers, ANSA supports layered and mixed-element meshing, which has been successfully tested in other configurations.
-The same methodology is also applicable to quasi-2D meshes in ANSA.
 In all cases, a fully curved volume mesh is generated using ANSA's internal second-order volume mesher.
 PyHOPE reads the exported CGNS file, reconstructs boundary conditions, and creates a fully curved HOPR mesh including connectivity information for vertices, edges, and sides.
 The final mesh contains 147763 second-order elements and can be processed by PyHOPE on any standard machine using approximately 520 MB RAM (peak RSS).
-Notably, this workflow enables full utilization of CFD meshes generated by commercial tools such as ANSA and Pointwise, which are widely used in industry.
+Notably, this workflow enables full utilization of CFD meshes generated by commercial tools such as ANSA and Pointwise [@pointwise2025], which are widely used in industry.
 \autoref{fig:CRM} shows the second-order unstructured surface grid and the instantaneous distribution of surface pressure on the NASA CRM, respectively.
 The latter was calculated using FLEXI, an open-source framework for the solution of the compressible Navier-Stokes equations using DGSEM.
 
