+++ paper.bib
@@ -1,3 +1,12 @@
+@manual{ansa2024,
+  title         = {ANSA Pre-processor},
+  author        = {{BETA CAE Systems S.A.}},
+  year          = 2024,
+  organization  = {BETA CAE Systems S.A.},
+  address       = {Thessaloniki, Greece},
+  note          = {Version 2024.1},
+  url           = {https://www.beta-cae.com/ansa.htm},
+}
 @article{Fasoulas2019,
   author        = {Fasoulas, S. and Munz, C.-D. and Pfeiffer, M. and Beyer, J. and Binder, T. and Copplestone, S. and Mirza, A. and Nizenkov, P. and Ortwein, P. and Reschke, W.},
   journal       = {Physics of Fluids},
@@ -36,6 +45,12 @@
   abstract      = {Gmsh is an open-source 3-D finite element grid generator with a build-in CAD engine and post-processor. Its design goal is to provide a fast, light and user-friendly meshing tool with parametric input and advanced visualization capabilities. This paper presents the overall philosophy, the main design choices and some of the original algorithms implemented in Gmsh.},
   doi           = {10.1002/nme.2579},
   publisher     = {Wiley},
+}
+@misc{hdf5,
+  author        = {{The HDF Group}},
+  title         = {Hierarchical Data Format, version 5},
+  year          = {1997-2022},
+  howpublished  = {\url{https://www.hdfgroup.org/HDF5}},
 }
 @phdthesis{Hindenlang2014,
   author        = {Hindenlang, Florian},
@@ -83,9 +98,25 @@
   doi           = {10.1016/j.cpc.2024.109388},
   publisher     = {Elsevier BV},
 }
+@manual{pointwise2025,
+  title         = {Pointwise Mesh Generation Software},
+  author        = {{Pointwise, Inc.}},
+  year          = 2025,
+  url           = {https://www.pointwise.com},
+}
 @misc{RuedaRamirez2017,
   author        = {Rueda-Ramirez, A. and Schlottke-Lakemper, M. and Gassner, G. J. and Astanin, A. and Winters, A. R.},
   title         = {{DGSEM} for general advection-diffusion equations},
   url           = {https://github.com/project-fluxo/fluxo},
   year          = 2017,
+}
+@inproceedings{Rumsey2012,
+  author        = {Rumsey, Christopher and Wedan, Bruce and Hauser, Thomas and Poinot, Marc},
+  booktitle     = {50th AIAA Aerospace Sciences Meeting including the New Horizons Forum and Aerospace Exposition},
+  title         = {Recent Updates to the CFD General Notation System (CGNS)},
+  year          = 2012,
+  month         = jan,
+  publisher     = {American Institute of Aeronautics and Astronautics},
+  abstract      = {The CFD General Notation System (CGNS)--a general, portable, and extensible standard for the storage and retrieval of computational fluid dynamics (CFD) analysis data--has been in existence for more than a decade (Version 1.0 was released in May 1998). Both structured and unstructured CFD data are covered by the standard, and CGNS can be easily extended to cover any sort of data imaginable, while retaining backward compatibility with existing CGNS data files and software. Although originally designed for CFD, it is readily extendable to any field of computational analysis. In early 2011, CGNS Version 3.1 was released, which added significant capabilities. This paper describes these recent enhancements and highlights the continued usefulness of the CGNS methodology},
+  doi           = {10.2514/6.2012-1264},
 }
