### 2025-10-10 4fe0f15
DebugMesh: Add IJK sorting
