### 2025-10-15 f58cd0a
FEMconnect: Correctly consider multiple periodicity
