+++ pyhope/mesh/fem/fem.py
@@ -26,7 +26,6 @@
 # Standard libraries
 # ----------------------------------------------------------------------------------------------------------------------------------
 from collections import defaultdict
-from itertools import chain
 from typing import Dict, Tuple, cast
 # ----------------------------------------------------------------------------------------------------------------------------------
 # Third-party libraries
@@ -41,13 +40,13 @@ import numpy as np
 # ==================================================================================================================================
 
 
-def copysign_int(x: int, y: int) -> int:
-    """ Return a int with the magnitude (absolute value) of x but the sign of y
-    """
-    # Standard libraries -----------------------------------
-    import math
-    # ------------------------------------------------------
-    return int(math.copysign(x, y))
+# def copysign_int(x: int, y: int) -> int:
+#     """ Return a int with the magnitude (absolute value) of x but the sign of y
+#     """
+#     # Standard libraries -----------------------------------
+#     import math
+#     # ------------------------------------------------------
+#     return int(math.copysign(x, y))
 
 
 def FEMConnect() -> None:
@@ -75,38 +74,196 @@ def FEMConnect() -> None:
     elems     = mesh_vars.elems
     periNodes = mesh_vars.periNodes
 
-    # Create a bidirectional lookup using a single dictionary comprehension with chain
-    periDict = { k: v for k, v in chain(((int(node), int(peri)) for (node, _), peri in periNodes.items()),
-                                        ((int(peri), int(node)) for (node, _), peri in periNodes.items()))}
+    # Build a graph of all periodic connections across all BCs
+    periGraph: dict[int, set[int]] = defaultdict(set)
+    for (node, _bc), peri in periNodes.items():
+        node, peri = int(node), int(peri)
+        periGraph[node].add(peri)
+        periGraph[peri].add(node)
 
-    # Build mapping of each node -> set of element indices that include that node.
+    # Find connected components; representative = min(node indices in component)
+    nodeFirst: dict[int, int]      = {}  # node -> representative (canonical)
+    nodeGroup: dict[int, set[int]] = {}  # representative -> full member set
+
+    for start in list(periGraph.keys()):
+        if start in nodeFirst:
+            continue
+
+        stack = [start]
+        comp  = set()
+
+        while stack:
+            cur = stack.pop()
+            if cur in comp:
+                continue
+
+            comp.add(cur)
+            for nxt in periGraph[cur]:
+                if nxt not in comp:
+                    stack.append(nxt)
+
+        rep = min(comp)
+        for v in comp:
+            nodeFirst[v] = rep
+
+        nodeGroup[rep] = comp
+
+    # Convenience mapping: each node -> full set of periodic equivalents (including itself)
+    periGroups: dict[int, set[int]] = {}
+    for rep, members in nodeGroup.items():
+        for v in members:
+            periGroups[v] = members
+
+    # Build mapping of each node -> set of element indices that include that node
     nodeToElements = defaultdict(set)
     for idx, elem in enumerate(elems):
-        for n in cast(np.ndarray, elem.nodes)[:cast(int, elem.type) % 10]:
-            nodeToElements[int(n)].add(idx)
+        for node in cast(np.ndarray, elem.nodes)[:cast(int, elem.type) % 10]:
+            nodeToElements[int(node)].add(idx)
 
     # Precompute combined connectivity for each node
     # > For a given node, the combined set is:
[…]
