### 2025-10-15 287170c
Linting: Update to ty 0.0.1a22
