### 2025-10-15 e6adb67
Readintools: Explicitly cast to str before strToFloatOrPi
