### 2025-10-15 09ef9f3
JOSS Manuscript: Minor fixes
