### 2025-10-21 7c085a0
Readintools: Explicitly cast default parameter to str
