### 2025-10-21 b8cbe05
Documentation: Update mesh sorting documentation
