### 2025-10-22 #8769 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
