### 2025-10-27 b084d3a
minor changes in paper.md
