+++ paper.md
@@ -44,18 +44,18 @@ bibliography: paper.bib
 
 # Summary
 PyHOPE (Python High-Order Preprocessing Environment) is a framework for generating and converting high-order meshes composed of standard 3D element types, designed for massively parallel spectral element solvers on high-performance computing (HPC) systems.
-PyHOPE builds on and extends Gmsh [@Geuzaine2009] which is used for the initial mesh generation and/or mesh read-in before conversion of the mesh to its internal representation and application of boundary conditions.
-Parallel read-in is crucial on HPC clusters which typically use parallel distributed file systems to enable and scale storage access by striping data across multiple servers.
-Primary output format of PyHOPE is the HOPR [@Hindenlang2014] HDF5 curved mesh format which is specifically designed for parallel read-in of unstructured three-dimensional meshes of arbitrary order, including tetrahedra, pyramids, prisms, and hexahedra.
+PyHOPE builds on and extends Gmsh [@Geuzaine2009], which is used for the initial mesh generation and/or mesh read-in before conversion of the mesh to its internal representation and application of boundary conditions.
+Parallel read-in is crucial on HPC clusters, which typically use parallel distributed file systems to enable and scale storage access by striping data across multiple servers.
+The primary output format of PyHOPE is the HOPR [@Hindenlang2014] HDF5 curved mesh format, which is specifically designed for parallel read-in of unstructured three-dimensional meshes of arbitrary order, including tetrahedra, pyramids, prisms, and hexahedra.
 Information stored in HOPR format facilitates non-overlapping input/output (I/O) through collocation of the required mesh information, including the vertex and side information together with element connectivity, in per-element packages.
 Each package is assigned a unique identifier via ordering along structured dimensions or a space-filling curve.
 
 # Statement of Need
 The Discontinuous Galerkin Spectral Element Method (DGSEM) is a powerful numerical approach for solving partial differential equations, particularly in high-performance computing applications.
-Prominent examples of DGSEM codes originating and being actively developed at the University of Stuttgart include the FLEXI family[^1] geared towards solving the compressible (multiphase) Navier-Stokes equations along with PICLas[^2] which focuses on plasma simulation with a Particle-in-Cell/Direct Simulation Monte Carlo approach.
+Prominent examples of DGSEM codes originating and being actively developed at the University of Stuttgart include the FLEXI family[^1] geared towards solving the compressible (multiphase) Navier-Stokes equations, along with PICLas[^2], which focuses on plasma simulation with a Particle-in-Cell/Direct Simulation Monte Carlo approach.
 A crucial aspect of DGSEM is the mapping of the equations to be solved from reference space to physical space, which relies on the computation of the Jacobian determinant to ensure accurate transformations for curved high-order elements.
 At the same time, DGSEM features a highly local stencil since grid elements are connected solely via the numerical flux through adjacent element faces.
-Providing mesh and solution data in an on-disc data format which facilitates non-overlapping I/O is key to efficient parallel data access, thereby minimizing execution time.
+Providing mesh and solution data in an on-disc data format, which facilitates non-overlapping I/O, is key to efficient parallel data access, thereby minimizing execution time.
 While HOPR has traditionally served as the reference implementation of a mesh generator for this process, PyHOPE is a modern alternative that enhances readability and extensibility.
 Designed with a clear code structure and modularization in mind, PyHOPE offers a more user-friendly and adaptable solution for researchers and engineers working on high-order mesh generation and transformation.
 
@@ -63,7 +63,7 @@ Designed with a clear code structure and modularization in mind, PyHOPE offers a
 [^2]: [https://github.com/piclas-framework/piclas](https://github.com/piclas-framework/piclas)
 
 # State of the Field
[…]
