### 2025-10-28 #8769 @danielskatz
@kopperp - please check this proof and confirm everything looks ok
