### 2025-10-29 8ecb140
JOSS Manuscript: Minor fixes
