### 2025-11-03 #8769 @kopperp
> Can you explain what you refer to? I don't see an attempt to include your email in the XML, just the JOSS one, and we don't list author emails in papers as far as I know.

You are right, I looked at the [jats file](https://github.com/openjournals/joss-papers/pull/7361/files#diff-1b47cb9e40a0e9358c8c8f67ea25c1ae09da02557b2b79cdc77b291172e10754R102) by mistake. Everything looks good to me.
